@@ -4,7 +4,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Data.Common;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
@@ -15,27 +14,27 @@ using System.Security;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Transactions;
 using Microsoft.Data.Common;
+using Microsoft.Data.Common.ConnectionString;
 using Microsoft.Data.ProviderBase;
 using Microsoft.Data.SqlClient.ConnectionPool;
 using Microsoft.Identity.Client;
-using System.Transactions;
-
 
 namespace Microsoft.Data.SqlClient
 {
-    internal class SessionStateRecord
+    internal sealed class SessionStateRecord
     {
         internal bool _recoverable;
-        internal UInt32 _version;
-        internal Int32 _dataLength;
+        internal uint _version;
+        internal int _dataLength;
         internal byte[] _data;
     }
 
-    internal class SessionData
+    internal sealed class SessionData
     {
         internal const int _maxNumberOfSessionStates = 256;
-        internal UInt32 _tdsVersion;
+        internal uint _tdsVersion;
         internal bool _encrypted;
 
         internal string _database;
@@ -104,17 +103,17 @@ namespace Microsoft.Data.SqlClient
         }
     }
 
-    sealed internal class SqlInternalConnectionTds : SqlInternalConnection, IDisposable
+    internal sealed class SqlInternalConnectionTds : SqlInternalConnection, IDisposable
     {
         // https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/wiki/retry-after#simple-retry-for-errors-with-http-error-codes-500-600
         internal const int MsalHttpRetryStatusCode = 429;
-
         // Connection re-route limit
-        internal const int _maxNumberOfRedirectRoute = 10;
+        internal const int MaxNumberOfRedirectRoute = 10;
 
         // CONNECTION AND STATE VARIABLES
         private readonly SqlConnectionPoolGroupProviderInfo _poolGroupProviderInfo; // will only be null when called for ChangePassword, or creating SSE User Instance
         private TdsParser _parser;
+
         private SqlLoginAck _loginAck;
         private SqlCredential _credential;
         private FederatedAuthenticationFeatureExtensionData _fedAuthFeatureExtensionData;
@@ -128,7 +127,6 @@ namespace Microsoft.Data.SqlClient
         // Federated Authentication
         // Response obtained from the server for FEDAUTHREQUIRED prelogin option.
         internal bool _fedAuthRequired;
-
         internal bool _federatedAuthenticationRequested;
         internal bool _federatedAuthenticationAcknowledged;
         internal bool _federatedAuthenticationInfoRequested; // Keep this distinct from _federatedAuthenticationRequested, since some fedauth library types may not need more info
@@ -137,7 +135,7 @@ namespace Microsoft.Data.SqlClient
         // The Federated Authentication returned by TryGetFedAuthTokenLocked or GetFedAuthToken.
         SqlFedAuthToken _fedAuthToken = null;
         internal byte[] _accessTokenInBytes;
-        internal readonly Func<SqlAuthenticationParameters, CancellationToken,Task<SqlAuthenticationToken>> _accessTokenCallback;
+        internal readonly Func<SqlAuthenticationParameters, CancellationToken, Task<SqlAuthenticationToken>> _accessTokenCallback;
 
         private readonly ActiveDirectoryAuthenticationTimeoutRetryHelper _activeDirectoryAuthTimeoutRetryHelper;
 
@@ -157,7 +155,7 @@ namespace Microsoft.Data.SqlClient
         }
 
         /// <summary>
-        /// Get or set if SQLDNSCaching FeatureExtAck is supported by the server.
+        /// Get or set if SQLDNSCaching is supported by the server.
         /// </summary>
         internal bool IsSQLDNSCachingSupported
         {
@@ -188,20 +186,20 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        private bool DNSCachingBeforeRedirect = false;
+        private bool _dnsCachingBeforeRedirect = false;
 
         /// <summary>
-        /// Get or set if the control ring send redirect token and SQLDNSCaching FeatureExtAck with true
+        /// Get or set if the control ring send redirect token and feature ext ack with true for DNSCaching
         /// </summary>
         internal bool IsDNSCachingBeforeRedirectSupported
         {
             get
             {
-                return DNSCachingBeforeRedirect;
+                return _dnsCachingBeforeRedirect;
             }
             set
             {
-                DNSCachingBeforeRedirect = value;
+                _dnsCachingBeforeRedirect = value;
             }
         }
 
@@ -214,7 +212,7 @@ namespace Microsoft.Data.SqlClient
         internal byte _tceVersionSupported;
 
         // The pool that this connection is associated with, if at all it is.
-        private DbConnectionPool _dbConnectionPool;
+        private IDbConnectionPool _dbConnectionPool;
 
         // This is used to preserve the authentication context object if we decide to cache it for subsequent connections in the same pool.
         // This will finally end up in _dbConnectionPool.AuthenticationContexts, but only after 1 successful login to SQL Server using this context.
@@ -253,9 +251,54 @@ namespace Microsoft.Data.SqlClient
         // If the context is expiring within the next 10 mins, then create a new context, irrespective of if another thread is trying to do the same.
         private static readonly TimeSpan _dbAuthenticationContextUnLockedRefreshTimeSpan = new TimeSpan(hours: 0, minutes: 10, seconds: 00);
 
-        private readonly TimeoutTimer _timeout;
+        // The errors in the transient error set are contained in
+        // https://azure.microsoft.com/en-us/documentation/articles/sql-database-develop-error-messages/#transient-faults-connection-loss-and-other-temporary-errors
+        private static readonly HashSet<int> s_transientErrors = new HashSet<int>
+            {
+            // SQL Error Code: 4060
+            // Cannot open database "%.*ls" requested by the login. The login failed.
+            4060,
 
-        private static HashSet<int> transientErrors = new HashSet<int>();
+            // SQL Error Code: 10928
+            // Resource ID: %d. The %s limit for the database is %d and has been reached.
+            10928,
+
+            // SQL Error Code: 10929
+            // Resource ID: %d. The %s minimum guarantee is %d, maximum limit is %d and the current usage for the database is %d.
+            // However, the server is currently too busy to support requests greater than %d for this database.
+            10929,
+
+            // SQL Error Code: 40197
+            // You will receive this error, when the service is down due to software or hardware upgrades, hardware failures,
+            // or any other failover problems. The error code (%d) embedded within the message of error 40197 provides
+            // additional information about the kind of failure or failover that occurred. Some examples of the error codes are
+            // embedded within the message of error 40197 are 40020, 40143, 40166, and 40540.
+            40197,
+            40020,
+            40143,
+            40166,
+
+            // The service has encountered an error processing your request. Please try again.
+            40540,
+
+            // The service is currently busy. Retry the request after 10 seconds. Incident ID: %ls. Code: %d.
+            40501,
+
+            // Database '%.*ls' on server '%.*ls' is not currently available. Please retry the connection later.
+            // If the problem persists, contact customer support, and provide them the session tracing ID of '%.*ls'.
+            40613,
+
+            // Can not connect to the SQL pool since it is paused. Please resume the SQL pool and try again.
+            42108,
+
+            // The SQL pool is warming up. Please try again.
+            42109
+
+            // Do federation errors deserve to be here ?
+            // Note: Federation errors 10053 and 10054 might also deserve inclusion in your retry logic.
+            // 10053
+            // 10054
+        };
 
         internal SessionData CurrentSessionData
         {
@@ -294,28 +337,28 @@ namespace Microsoft.Data.SqlClient
         // 2. _parserLock will also be taken during close (to prevent closing in the middle of a write)
         // 3. Whenever you have the _parserLock and are calling a method that would cause the connection to close if it failed (with the exception of any writing method), you MUST set ThreadHasParserLockForClose to true
         //      * This is to prevent the connection deadlocking with itself (since you already have the _parserLock, and Closing the connection will attempt to re-take that lock)
-        //      * It is safe to set ThreadHasParserLockForClose to true when writing as well, but it is unneccesary
+        //      * It is safe to set ThreadHasParserLockForClose to true when writing as well, but it is unnecessary
         //      * If you have a method that takes _parserLock, it is a good idea check ThreadHasParserLockForClose first (if you don't expect _parserLock to be taken by something higher on the stack, then you should at least assert that it is false)
         // 4. ThreadHasParserLockForClose is thread-specific - this means that you must set it to false before returning a Task, and set it back to true in the continuation
         // 5. ThreadHasParserLockForClose should only be modified if you currently own the _parserLock
         // 6. Reading ThreadHasParserLockForClose is thread-safe
         internal class SyncAsyncLock
         {
-            SemaphoreSlim semaphore = new SemaphoreSlim(1);
+            private SemaphoreSlim _semaphore = new SemaphoreSlim(1);
 
             internal void Wait(bool canReleaseFromAnyThread)
             {
-                Monitor.Enter(semaphore); // semaphore is used as lock object, no relation to SemaphoreSlim.Wait/Release methods
-                if (canReleaseFromAnyThread || semaphore.CurrentCount == 0)
+                Monitor.Enter(_semaphore); // semaphore is used as lock object, no relation to SemaphoreSlim.Wait/Release methods
+                if (canReleaseFromAnyThread || _semaphore.CurrentCount == 0)
                 {
-                    semaphore.Wait();
+                    _semaphore.Wait();
                     if (canReleaseFromAnyThread)
                     {
-                        Monitor.Exit(semaphore);
+                        Monitor.Exit(_semaphore);
                     }
                     else
                     {
-                        semaphore.Release();
+                        _semaphore.Release();
                     }
                 }
             }
@@ -326,21 +369,21 @@ namespace Microsoft.Data.SqlClient
                 bool hasMonitor = false;
                 try
                 {
-                    Monitor.TryEnter(semaphore, timeout, ref hasMonitor); // semaphore is used as lock object, no relation to SemaphoreSlim.Wait/Release methods
+                    Monitor.TryEnter(_semaphore, timeout, ref hasMonitor); // semaphore is used as lock object, no relation to SemaphoreSlim.Wait/Release methods
                     if (hasMonitor)
                     {
-                        if ((canReleaseFromAnyThread) || (semaphore.CurrentCount == 0))
+                        if ((canReleaseFromAnyThread) || (_semaphore.CurrentCount == 0))
                         {
-                            if (semaphore.Wait(timeout))
+                            if (_semaphore.Wait(timeout))
                             {
                                 if (canReleaseFromAnyThread)
                                 {
-                                    Monitor.Exit(semaphore);
+                                    Monitor.Exit(_semaphore);
                                     hasMonitor = false;
                                 }
                                 else
                                 {
-                                    semaphore.Release();
+                                    _semaphore.Release();
                                 }
                                 lockTaken = true;
                             }
@@ -355,20 +398,20 @@ namespace Microsoft.Data.SqlClient
                 {
                     if ((!lockTaken) && (hasMonitor))
                     {
-                        Monitor.Exit(semaphore);
+                        Monitor.Exit(_semaphore);
                     }
                 }
             }
 
             internal void Release()
             {
-                if (semaphore.CurrentCount == 0)
+                if (_semaphore.CurrentCount == 0)
                 {  //  semaphore methods were used for locking
-                    semaphore.Release();
+                    _semaphore.Release();
                 }
                 else
                 {
-                    Monitor.Exit(semaphore);
+                    Monitor.Exit(_semaphore);
                 }
             }
 
@@ -377,14 +420,14 @@ namespace Microsoft.Data.SqlClient
             {
                 get
                 {
-                    return semaphore.CurrentCount == 0;
+                    return _semaphore.CurrentCount == 0;
                 }
             }
 
-            // Necessary but not sufficient condition for thread to have lock (since sempahore may be obtained by any thread)
+            // Necessary but not sufficient condition for thread to have lock (since semaphore may be obtained by any thread)
             internal bool ThreadMayHaveLock()
             {
-                return Monitor.IsEntered(semaphore) || semaphore.CurrentCount == 0;
+                return Monitor.IsEntered(_semaphore) || _semaphore.CurrentCount == 0;
             }
         }
 
@@ -400,20 +443,14 @@ namespace Microsoft.Data.SqlClient
         }
 
         // OTHER STATE VARIABLES AND REFERENCES
-
         internal Guid _clientConnectionId = Guid.Empty;
 
         // Routing information (ROR)
-        RoutingInfo _routingInfo = null;
         private Guid _originalClientConnectionId = Guid.Empty;
         private string _routingDestination = null;
+        private readonly TimeoutTimer _timeout;
 
-        static SqlInternalConnectionTds()
-        {
-            populateTransientErrors();
-        }
-
-        // although the new password is generally not used it must be passed to the c'tor
+        // although the new password is generally not used it must be passed to the ctor
         // the new Login7 packet will always write out the new password (or a length of zero and no bytes if not present)
         //
         internal SqlInternalConnectionTds(
@@ -426,20 +463,20 @@ namespace Microsoft.Data.SqlClient
                 bool redirectedUserInstance,
                 SqlConnectionString userConnectionOptions = null, // NOTE: userConnectionOptions may be different to connectionOptions if the connection string has been expanded (see SqlConnectionString.Expand)
                 SessionData reconnectSessionData = null,
-                DbConnectionPool pool = null,
-                string accessToken = null,
                 bool applyTransientFaultHandling = false,
+                string accessToken = null,
+                IDbConnectionPool pool = null,
                 Func<SqlAuthenticationParameters, CancellationToken,
                 Task<SqlAuthenticationToken>> accessTokenCallback = null) : base(connectionOptions)
         {
-
 #if DEBUG
             if (reconnectSessionData != null)
             {
                 reconnectSessionData._debugReconnectDataApplied = true;
             }
             try
-            { // use this to help validate this object is only created after the following permission has been previously demanded in the current codepath
+            {
+                // use this to help validate this object is only created after the following permission has been previously demanded in the current codepath
                 if (userConnectionOptions != null)
                 {
                     // As mentioned above, userConnectionOptions may be different to connectionOptions, so we need to demand on the correct connection string
@@ -509,12 +546,13 @@ namespace Microsoft.Data.SqlClient
 
             _parserLock.Wait(canReleaseFromAnyThread: false);
             ThreadHasParserLockForClose = true;   // In case of error, let ourselves know that we already own the parser lock
+
             RuntimeHelpers.PrepareConstrainedRegions();
             try
             {
                 _timeout = TimeoutTimer.StartSecondsTimeout(connectionOptions.ConnectTimeout);
 
-                // If transient fault handling is enabled then we can retry the login upto the ConnectRetryCount.
+                // If transient fault handling is enabled then we can retry the login up to the ConnectRetryCount.
                 int connectionEstablishCount = applyTransientFaultHandling ? connectionOptions.ConnectRetryCount + 1 : 1;
                 int transientRetryIntervalInMilliSeconds = connectionOptions.ConnectRetryInterval * 1000; // Max value of transientRetryInterval is 60*1000 ms. The max value allowed for ConnectRetryInterval is 60
                 for (int i = 0; i < connectionEstablishCount; i++)
@@ -564,49 +602,6 @@ namespace Microsoft.Data.SqlClient
             SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.ctor|ADV> {0}, constructed new TDS internal connection", ObjectID);
         }
 
-        // The errors in the transient error set are contained in
-        // https://azure.microsoft.com/en-us/documentation/articles/sql-database-develop-error-messages/#transient-faults-connection-loss-and-other-temporary-errors
-        private static void populateTransientErrors()
-        {
-            // SQL Error Code: 4060
-            // Cannot open database "%.*ls" requested by the login. The login failed.
-            transientErrors.Add(4060);
-            // SQL Error Code: 10928
-            // Resource ID: %d. The %s limit for the database is %d and has been reached.
-            transientErrors.Add(10928);
-            // SQL Error Code: 10929
-            // Resource ID: %d. The %s minimum guarantee is %d, maximum limit is %d and the current usage for the database is %d.
-            // However, the server is currently too busy to support requests greater than %d for this database.
-            transientErrors.Add(10929);
-            // SQL Error Code: 40197
-            // You will receive this error, when the service is down due to software or hardware upgrades, hardware failures,
-            // or any other failover problems. The error code (%d) embedded within the message of error 40197 provides
-            // additional information about the kind of failure or failover that occurred. Some examples of the error codes are
-            // embedded within the message of error 40197 are 40020, 40143, 40166, and 40540.
-            transientErrors.Add(40197);
-            transientErrors.Add(40020);
-            transientErrors.Add(40143);
-            transientErrors.Add(40166);
-            // The service has encountered an error processing your request. Please try again.
-            transientErrors.Add(40540);
-            // The service is currently busy. Retry the request after 10 seconds. Incident ID: %ls. Code: %d.
-            transientErrors.Add(40501);
-            // Database '%.*ls' on server '%.*ls' is not currently available. Please retry the connection later.
-            // If the problem persists, contact customer support, and provide them the session tracing ID of '%.*ls'.
-            transientErrors.Add(40613);
-
-            // Can not connect to the SQL pool since it is paused. Please resume the SQL pool and try again.
-            transientErrors.Add(42108);
-
-            // The SQL pool is warming up. Please try again.
-            transientErrors.Add(42109);
-            // Do federation errors deserve to be here ?
-            // Note: Federation errors 10053 and 10054 might also deserve inclusion in your retry logic.
-            //transientErrors.Add(10053);
-            //transientErrors.Add(10054);
-        }
-
-
         // Returns true if the Sql error is a transient.
         private bool IsTransientError(SqlException exc)
         {
@@ -616,7 +611,7 @@ namespace Microsoft.Data.SqlClient
             }
             foreach (SqlError error in exc.Errors)
             {
-                if (transientErrors.Contains(error.Number))
+                if (s_transientErrors.Contains(error.Number))
                 {
                     // When server timeouts, connection is doomed. Reset here to allow reconnect.
                     UnDoomThisConnection();
@@ -650,15 +645,9 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        internal RoutingInfo RoutingInfo
-        {
-            get
-            {
-                return _routingInfo;
-            }
-        }
+        internal RoutingInfo RoutingInfo { get; private set; } = null;
 
-        override internal SqlInternalTransaction CurrentTransaction
+        internal override SqlInternalTransaction CurrentTransaction
         {
             get
             {
@@ -666,7 +655,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        override internal SqlInternalTransaction AvailableInternalTransaction
+        internal override SqlInternalTransaction AvailableInternalTransaction
         {
             get
             {
@@ -674,8 +663,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-
-        override internal SqlInternalTransaction PendingTransaction
+        internal override SqlInternalTransaction PendingTransaction
         {
             get
             {
@@ -699,7 +687,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        override internal bool IsLockedForBulkCopy
+        internal override bool IsLockedForBulkCopy
         {
             get
             {
@@ -707,7 +695,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        override protected internal bool IsNonPoolableTransactionRoot
+        protected internal override bool IsNonPoolableTransactionRoot
         {
             get
             {
@@ -715,7 +703,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        override internal bool Is2008OrNewer
+        internal override bool Is2008OrNewer
         {
             get
             {
@@ -755,7 +743,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        override protected bool ReadyToPrepareTransaction
+        protected override bool ReadyToPrepareTransaction
         {
             get
             {
@@ -765,7 +753,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        override public string ServerVersion
+        public override string ServerVersion
         {
             get
             {
@@ -773,6 +761,7 @@ namespace Microsoft.Data.SqlClient
                        (short)_loginAck.minorVersion, _loginAck.buildNum));
             }
         }
+
         public int ServerProcessId
         {
             get
@@ -808,16 +797,16 @@ namespace Microsoft.Data.SqlClient
         // GENERAL METHODS
         ////////////////////////////////////////////////////////////////////////////////////////
         [SuppressMessage("Microsoft.Globalization", "CA1303:DoNotPassLiteralsAsLocalizedParameters")] // copied from Triaged.cs
-        override protected void ChangeDatabaseInternal(string database)
+        protected override void ChangeDatabaseInternal(string database)
         {
             // MDAC 73598 - add brackets around database
             database = SqlConnection.FixupDatabaseTransactionName(database);
-            System.Threading.Tasks.Task executeTask = _parser.TdsExecuteSQLBatch("use " + database, ConnectionOptions.ConnectTimeout, null, _parser._physicalStateObj, sync: true);
+            Task executeTask = _parser.TdsExecuteSQLBatch("use " + database, ConnectionOptions.ConnectTimeout, null, _parser._physicalStateObj, sync: true);
             Debug.Assert(executeTask == null, "Shouldn't get a task when doing sync writes");
             _parser.Run(RunBehavior.UntilDone, null, null, null, _parser._physicalStateObj);
         }
 
-        override public void Dispose()
+        public override void Dispose()
         {
             SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.Dispose|ADV> {0} disposing", ObjectID);
             try
@@ -831,7 +820,7 @@ namespace Microsoft.Data.SqlClient
                 }
             }
             finally
-            { // UNDONE: MDAC 77928
+            {
                 // close will always close, even if exception is thrown
                 // remember to null out any object references
                 _loginAck = null;
@@ -840,7 +829,7 @@ namespace Microsoft.Data.SqlClient
             base.Dispose();
         }
 
-        override internal void ValidateConnectionForExecute(SqlCommand command)
+        internal override void ValidateConnectionForExecute(SqlCommand command)
         {
             TdsParser parser = _parser;
             if ((parser == null) || (parser.State == TdsParserState.Broken) || (parser.State == TdsParserState.Closed))
@@ -901,7 +890,7 @@ namespace Microsoft.Data.SqlClient
         /// <para>
         /// When using Implicit transaction unbinding,
         /// verify that the enlisted transaction is active.
-        /// If it is not active, and the transaction object has been diposed, unbind from the transaction.
+        /// If it is not active, and the transaction object has been disposed, unbind from the transaction.
         /// If it is not active and not disposed, throw an exception.
         /// </para>
         /// </remarks>
@@ -948,7 +937,7 @@ namespace Microsoft.Data.SqlClient
         // POOLING METHODS
         ////////////////////////////////////////////////////////////////////////////////////////
 
-        override protected void Activate(Transaction transaction)
+        protected override void Activate(Transaction transaction)
         {
             FailoverPermissionDemand(); // Demand for unspecified failover pooled connections
 
@@ -973,7 +962,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        override protected void InternalDeactivate()
+        protected override void InternalDeactivate()
         {
             // When we're deactivated, the user must have called End on all
             // the async commands, or we don't know that we're in a state that
@@ -994,7 +983,6 @@ namespace Microsoft.Data.SqlClient
                 Debug.Assert(_parser != null || IsConnectionDoomed, "Deactivating a disposed connection?");
                 if (_parser != null)
                 {
-
                     _parser.Deactivate(IsConnectionDoomed);
 
                     if (!IsConnectionDoomed)
@@ -1010,7 +998,7 @@ namespace Microsoft.Data.SqlClient
         {
             // For implicit pooled connections, if connection reset behavior is specified,
             // reset the database and language properties back to default.  It is important
-            // to do this on activate so that the hashtable is correct before SqlConnection
+            // to do this on activate so that the dictionary is correct before SqlConnection
             // obtains a clone.
 
             Debug.Assert(!HasLocalTransactionFromAPI, "Upon ResetConnection SqlInternalConnectionTds has a currently ongoing local transaction.");
@@ -1024,7 +1012,7 @@ namespace Microsoft.Data.SqlClient
                 // https://github.com/dotnet/SqlClient/issues/2970
                 _parser.PrepareResetConnection(EnlistedTransaction is not null && Pool is not null);
 
-                // Reset hashtable values, since calling reset will not send us env_changes.
+                // Reset dictionary values, since calling reset will not send us env_changes.
                 CurrentDatabase = _originalDatabase;
                 _currentLanguage = _originalLanguage;
             }
@@ -1046,7 +1034,7 @@ namespace Microsoft.Data.SqlClient
         // LOCAL TRANSACTION METHODS
         ////////////////////////////////////////////////////////////////////////////////////////
 
-        override internal void DisconnectTransaction(SqlInternalTransaction internalTransaction)
+        internal override void DisconnectTransaction(SqlInternalTransaction internalTransaction)
         {
             TdsParser parser = Parser;
 
@@ -1061,7 +1049,7 @@ namespace Microsoft.Data.SqlClient
             ExecuteTransaction(transactionRequest, name, iso, null, false);
         }
 
-        override internal void ExecuteTransaction(TransactionRequest transactionRequest, string name, System.Data.IsolationLevel iso, SqlInternalTransaction internalTransaction, bool isDelegateControlRequest)
+        internal override void ExecuteTransaction(TransactionRequest transactionRequest, string name, System.Data.IsolationLevel iso, SqlInternalTransaction internalTransaction, bool isDelegateControlRequest)
         {
             if (IsConnectionDoomed)
             {  // doomed means we can't do anything else...
@@ -1182,19 +1170,17 @@ namespace Microsoft.Data.SqlClient
                     }
                 }
 
-
-
                 // SQLBUDT #20010853 - Promote, Commit and Rollback requests for
                 // delegated transactions often happen while there is an open result
                 // set, so we need to handle them by using a different MARS session,
                 // otherwise we'll write on the physical state objects while someone
                 // else is using it.  When we don't have MARS enabled, we need to
-                // lock the physical state object to syncronize it's use at least
+                // lock the physical state object to synchronize it's use at least
                 // until we increment the open results count.  Once it's been
                 // incremented the delegated transaction requests will fail, so they
                 // won't stomp on anything.
                 //
-                // We need to keep this lock through the duration of the TM reqeuest
+                // We need to keep this lock through the duration of the TM request
                 // so that we won't hijack a different request's data stream and a
                 // different request won't hijack ours, so we have a lock here on
                 // an object that the ExecTMReq will also lock, but since we're on
@@ -1217,7 +1203,7 @@ namespace Microsoft.Data.SqlClient
                     }
                 }
 
-                // SQLBU #406778 - _parser may be nulled out during TdsExecuteTrannsactionManagerRequest.
+                // SQLBU #406778 - _parser may be nulled out during TdsExecuteTransactionManagerRequest.
                 //  Only use local variable after this call.
                 _parser.TdsExecuteTransactionManagerRequest(null, requestType, transactionName, isoLevel,
                     ConnectionOptions.ConnectTimeout, internalTransaction, stateObj, isDelegateControlRequest);
@@ -1241,20 +1227,20 @@ namespace Microsoft.Data.SqlClient
         // DISTRIBUTED TRANSACTION METHODS
         ////////////////////////////////////////////////////////////////////////////////////////
 
-        override internal void DelegatedTransactionEnded()
+        internal override void DelegatedTransactionEnded()
         {
             // TODO: I don't like the way that this works, but I don't want to rototill the entire pooler to avoid this call.
             base.DelegatedTransactionEnded();
         }
 
-        override protected byte[] GetDTCAddress()
+        protected override byte[] GetDTCAddress()
         {
             byte[] dtcAddress = _parser.GetDTCAddress(ConnectionOptions.ConnectTimeout, _parser.GetSession(this));
             Debug.Assert(dtcAddress != null, "null dtcAddress?");
             return dtcAddress;
         }
 
-        override protected void PropagateTransactionCookie(byte[] cookie)
+        protected override void PropagateTransactionCookie(byte[] cookie)
         {
             _parser.PropagateDistributedTransaction(cookie, ConnectionOptions.ConnectTimeout, _parser._physicalStateObj);
         }
@@ -1267,7 +1253,7 @@ namespace Microsoft.Data.SqlClient
         {
             _parser.Run(RunBehavior.UntilDone, null, null, null, _parser._physicalStateObj);
 
-            if (_routingInfo == null)
+            if (RoutingInfo == null)
             {
                 // ROR should not affect state of connection recovery
                 if (_federatedAuthenticationRequested && !_federatedAuthenticationAcknowledged)
@@ -1319,12 +1305,13 @@ namespace Microsoft.Data.SqlClient
 
             // for non-pooled connections, enlist in a distributed transaction
             // if present - and user specified to enlist
-            if (enlistOK && ConnectionOptions.Enlist && _routingInfo == null)
+            if (enlistOK && ConnectionOptions.Enlist && RoutingInfo == null)
             {
                 _parser._physicalStateObj.SniContext = SniContext.Snix_AutoEnlist;
                 Transaction tx = ADP.GetCurrentTransaction();
                 Enlist(tx);
             }
+
             _parser._physicalStateObj.SniContext = SniContext.Snix_Login;
         }
 
@@ -1369,9 +1356,9 @@ namespace Microsoft.Data.SqlClient
 
             login.language = _currentLanguage;
             if (!login.userInstance)
-            { // Do not send attachdbfilename or database to SSE primary instance
+            {
+                // Do not send attachdbfilename or database to SSE primary instance
                 login.database = CurrentDatabase;
-                ;
                 login.attachDBFilename = ConnectionOptions.AttachDBFilename;
             }
 
@@ -1404,8 +1391,8 @@ namespace Microsoft.Data.SqlClient
             // in Login7, indicating the intent to use Active Directory Authentication for SQL Server.
             if (ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryPassword
                 || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryInteractive
-                || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryServicePrincipal
                 || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryDeviceCodeFlow
+                || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryServicePrincipal
                 || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryManagedIdentity
                 || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryMSI
                 || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryDefault
@@ -1438,10 +1425,8 @@ namespace Microsoft.Data.SqlClient
                 _federatedAuthenticationRequested = true;
             }
 
-            // The TCE, DATACLASSIFICATION and GLOBALTRANSACTIONS, UTF8 support feature are implicitly requested
-            requestedFeatures |= TdsEnums.FeatureExtension.Tce | TdsEnums.FeatureExtension.DataClassification | TdsEnums.FeatureExtension.GlobalTransactions;
-
-            requestedFeatures |= TdsEnums.FeatureExtension.UTF8Support;
+            // The GLOBALTRANSACTIONS, DATACLASSIFICATION, TCE, and UTF8 support features are implicitly requested
+            requestedFeatures |= TdsEnums.FeatureExtension.GlobalTransactions | TdsEnums.FeatureExtension.DataClassification | TdsEnums.FeatureExtension.Tce | TdsEnums.FeatureExtension.UTF8Support;
 
             // The AzureSQLSupport feature is implicitly set for ReadOnly login
             if (ConnectionOptions.ApplicationIntent == ApplicationIntent.ReadOnly)
@@ -1449,9 +1434,8 @@ namespace Microsoft.Data.SqlClient
                 requestedFeatures |= TdsEnums.FeatureExtension.AzureSQLSupport;
             }
 
-            // The SQLDNSCaching feature is implicitly set
+            // The SQLDNSCaching and JSON features are implicitly set
             requestedFeatures |= TdsEnums.FeatureExtension.SQLDNSCaching;
-
             requestedFeatures |= TdsEnums.FeatureExtension.JsonSupport;
 
             _parser.TdsLogin(login, requestedFeatures, _recoverySessionData, _fedAuthFeatureExtensionData, encrypt);
@@ -1467,14 +1451,16 @@ namespace Microsoft.Data.SqlClient
             // no-op, so no issues there.
             if (_parser != null)
             {
-
                 _parser.Disconnect();
             }
-            // TODO: Need a performance counter for Failed Connections
         }
 
-        private void OpenLoginEnlist(TimeoutTimer timeout, SqlConnectionString connectionOptions, SqlCredential credential,
-                    string newPassword, SecureString newSecurePassword, bool redirectedUserInstance)
+        private void OpenLoginEnlist(TimeoutTimer timeout,
+                                    SqlConnectionString connectionOptions,
+                                    SqlCredential credential,
+                                    string newPassword,
+                                    SecureString newSecurePassword,
+                                    bool redirectedUserInstance)
         {
             bool useFailoverPartner; // should we use primary or secondary first
             ServerInfo dataSource = new ServerInfo(connectionOptions);
@@ -1517,7 +1503,8 @@ namespace Microsoft.Data.SqlClient
                 else
                 {
                     _timeoutErrorInternal.SetFailoverScenario(false); // not a failover scenario
-                    LoginNoFailover(dataSource, 
+                    LoginNoFailover(
+                            dataSource, 
                             newPassword, 
                             newSecurePassword, 
                             redirectedUserInstance,
@@ -1546,7 +1533,6 @@ namespace Microsoft.Data.SqlClient
             }
             catch (Exception e)
             {
-                // UNDONE - should not be catching all exceptions!!!
                 if (ADP.IsCatchableExceptionType(e))
                 {
                     LoginFailure();
@@ -1564,11 +1550,10 @@ namespace Microsoft.Data.SqlClient
         //   to connect.
         private bool IsDoNotRetryConnectError(SqlException exc)
         {
-
             return (TdsEnums.LOGON_FAILED == exc.Number) // actual logon failed, i.e. bad password
                 || (TdsEnums.PASSWORD_EXPIRED == exc.Number) // actual logon failed, i.e. password isExpired
-                || (TdsEnums.IMPERSONATION_FAILED == exc.Number)  // Insuficient privelege for named pipe, among others
-                || exc._doNotReconnect; // Exception explicitly supressed reconnection attempts
+                || (TdsEnums.IMPERSONATION_FAILED == exc.Number)  // Insufficient privilege for named pipe, among others
+                || exc._doNotReconnect; // Exception explicitly suppressed reconnection attempts
         }
 
         // Attempt to login to a host that does not have a failover partner
@@ -1580,10 +1565,14 @@ namespace Microsoft.Data.SqlClient
         //  DEVNOTE: The logic in this method is paralleled by the logic in LoginWithFailover.
         //           Changes to either one should be examined to see if they need to be reflected in the other
         // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-        private void LoginNoFailover(ServerInfo serverInfo, string newPassword, SecureString newSecurePassword, bool redirectedUserInstance,
-                    SqlConnectionString connectionOptions, SqlCredential credential, TimeoutTimer timeout)
+        private void LoginNoFailover(ServerInfo serverInfo,
+                                     string newPassword,
+                                     SecureString newSecurePassword,
+                                     bool redirectedUserInstance,
+                                     SqlConnectionString connectionOptions,
+                                     SqlCredential credential,
+                                     TimeoutTimer timeout)
         {
-
             Debug.Assert(object.ReferenceEquals(connectionOptions, this.ConnectionOptions), "ConnectionOptions argument and property must be the same"); // consider removing the argument
             int routingAttempts = 0;
             ServerInfo originalServerInfo = serverInfo; // serverInfo may end up pointing to new object due to routing, original object is used to set CurrentDatasource
@@ -1594,16 +1583,16 @@ namespace Microsoft.Data.SqlClient
 
             ResolveExtendedServerName(serverInfo, !redirectedUserInstance, connectionOptions);
 
-            Boolean disableTnir = ShouldDisableTnir(connectionOptions);
+            bool disableTnir = ShouldDisableTnir(connectionOptions);
 
             long timeoutUnitInterval = 0;
 
-            Boolean isParallel = connectionOptions.MultiSubnetFailover || (connectionOptions.TransparentNetworkIPResolution && !disableTnir);
-
+            bool isParallel = connectionOptions.MultiSubnetFailover || (connectionOptions.TransparentNetworkIPResolution && !disableTnir);
 
             if (isParallel)
             {
                 float failoverTimeoutStep = connectionOptions.MultiSubnetFailover ? ADP.FailoverTimeoutStep : ADP.FailoverTimeoutStepForTnir;
+
                 // Determine unit interval
                 if (timeout.IsInfinite)
                 {
@@ -1625,11 +1614,10 @@ namespace Microsoft.Data.SqlClient
             //  back into the parser for the error cases.
             int attemptNumber = 0;
             TimeoutTimer intervalTimer = null;
-
             TimeoutTimer attemptOneLoginTimeout = timeout;
+
             while (true)
             {
-
                 Boolean isFirstTransparentAttempt = connectionOptions.TransparentNetworkIPResolution && !disableTnir && attemptNumber == 1;
 
                 if (isParallel)
@@ -1665,11 +1653,11 @@ namespace Microsoft.Data.SqlClient
                     _parser.Disconnect();
 
                 _parser = new TdsParser(ConnectionOptions.MARS, ConnectionOptions.Asynchronous);
-                Debug.Assert(SniContext.Undefined == Parser._physicalStateObj.SniContext, string.Format((IFormatProvider)null, "SniContext should be Undefined; actual Value: {0}", Parser._physicalStateObj.SniContext));
+                Debug.Assert(SniContext.Undefined == Parser._physicalStateObj.SniContext, $"SniContext should be Undefined; actual Value: {Parser._physicalStateObj.SniContext}");
 
                 try
                 {
-                    // UNDONE: ITEM12001110 (DB Mirroring Reconnect) Old behavior of not truly honoring timeout presevered
+                    // UNDONE: ITEM12001110 (DB Mirroring Reconnect) Old behavior of not truly honoring timeout preserved
                     //  for non-failover, non-MSF scenarios to avoid breaking changes as part of a QFE.  Consider fixing timeout
                     //  handling in next full release and removing ignoreSniOpenTimeout parameter.
 
@@ -1687,17 +1675,16 @@ namespace Microsoft.Data.SqlClient
 
                     if (connectionOptions.MultiSubnetFailover && ServerProvidedFailOverPartner != null)
                     {
-                        // connection succeeded: trigger exception if server sends failover partner and MultiSubnetFailover is used.
+                        // connection succeeded: trigger exception if server sends failover partner and MultiSubnetFailover is used
                         throw SQL.MultiSubnetFailoverWithFailoverPartner(serverProvidedFailoverPartner: true, internalConnection: this);
                     }
 
-                    if (_routingInfo != null)
+                    if (RoutingInfo != null)
                     {
                         SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.LoginNoFailover> Routed to {0}", serverInfo.ExtendedServerName);
-
-                        if (routingAttempts > _maxNumberOfRedirectRoute)
+                        if (routingAttempts > MaxNumberOfRedirectRoute)
                         {
-                            throw SQL.ROR_RecursiveRoutingNotSupported(this, _maxNumberOfRedirectRoute);
+                            throw SQL.ROR_RecursiveRoutingNotSupported(this, MaxNumberOfRedirectRoute);
                         }
 
                         if (timeout.IsExpired)
@@ -1705,7 +1692,7 @@ namespace Microsoft.Data.SqlClient
                             throw SQL.ROR_TimeoutAfterRoutingInfo(this);
                         }
 
-                        serverInfo = new ServerInfo(ConnectionOptions, _routingInfo, serverInfo.ResolvedServerName, serverInfo.ServerSPN);
+                        serverInfo = new ServerInfo(ConnectionOptions, RoutingInfo, serverInfo.ResolvedServerName, serverInfo.ServerSPN);
                         _timeoutErrorInternal.SetInternalSourceType(SqlConnectionInternalSourceType.RoutingDestination);
                         _originalClientConnectionId = _clientConnectionId;
                         _routingDestination = serverInfo.UserServerName;
@@ -1737,8 +1724,9 @@ namespace Microsoft.Data.SqlClient
                         || TdsParserState.Closed != _parser.State
                         || IsDoNotRetryConnectError(sqlex)
                         || timeout.IsExpired)
-                    {       // no more time to try again
-                        throw;  // Caller will call LoginFailure()
+                    {
+                        // no more time to try again
+                        throw; // Caller will call LoginFailure()
                     }
 
                     // Check sleep interval to make sure we won't exceed the timeout
@@ -1747,8 +1735,6 @@ namespace Microsoft.Data.SqlClient
                     {
                         throw;
                     }
-
-                    // TODO: Stash parser away somewhere so we can examine it's state during debugging
                 }
 
                 // We only get here when we failed to connect, but are going to re-try
@@ -1783,7 +1769,6 @@ namespace Microsoft.Data.SqlClient
                 // Sleep for a bit to prevent clogging the network with requests,
                 //  then update sleep interval for next iteration (max 1 second interval)
                 SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.LoginNoFailover|ADV> {0}, sleeping {1}[milisec]", ObjectID, sleepInterval);
-
                 Thread.Sleep(sleepInterval);
                 sleepInterval = (sleepInterval < 500) ? sleepInterval * 2 : 1000;
             }
@@ -1868,7 +1853,6 @@ namespace Microsoft.Data.SqlClient
                 TimeoutTimer timeout
             )
         {
-
             Debug.Assert(!connectionOptions.MultiSubnetFailover, "MultiSubnetFailover should not be set if failover partner is used");
             SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.LoginWithFailover|ADV> {0}, useFailover={1}[bool], primary={2}, failover={3}", ObjectID, useFailoverHost, primaryServerInfo.UserServerName, failoverHost);
 
@@ -1880,8 +1864,7 @@ namespace Microsoft.Data.SqlClient
 
             ResolveExtendedServerName(primaryServerInfo, !redirectedUserInstance, connectionOptions);
             if (ServerProvidedFailOverPartner == null)
-            {// No point in resolving the failover partner when we're going to override it below
-             // Don't resolve aliases if failover == primary // UNDONE: WHY?  Previous code in TdsParser.Connect did this, but the reason is not clear
+            {
                 ResolveExtendedServerName(failoverServerInfo, !redirectedUserInstance && failoverHost != primaryServerInfo.UserServerName, connectionOptions);
             }
 
@@ -1926,7 +1909,7 @@ namespace Microsoft.Data.SqlClient
                 }
 
                 _parser = new TdsParser(ConnectionOptions.MARS, ConnectionOptions.Asynchronous);
-                Debug.Assert(SniContext.Undefined == Parser._physicalStateObj.SniContext, string.Format((IFormatProvider)null, "SniContext should be Undefined; actual Value: {0}", Parser._physicalStateObj.SniContext));
+                Debug.Assert(SniContext.Undefined == Parser._physicalStateObj.SniContext, $"SniContext should be Undefined; actual Value: {Parser._physicalStateObj.SniContext}");
 
                 ServerInfo currentServerInfo;
                 if (useFailoverHost)
@@ -1964,33 +1947,34 @@ namespace Microsoft.Data.SqlClient
                             );
 
                     int routingAttempts = 0;
-                    while (_routingInfo != null)
+                    while (RoutingInfo != null)
                     {
-                        if (routingAttempts > _maxNumberOfRedirectRoute)
+                        if (routingAttempts > MaxNumberOfRedirectRoute)
                         {
-                            throw SQL.ROR_RecursiveRoutingNotSupported(this, _maxNumberOfRedirectRoute);
+                            throw SQL.ROR_RecursiveRoutingNotSupported(this, MaxNumberOfRedirectRoute);
                         }
                         routingAttempts++;
 
-                        SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.LoginWithFailover> Routed to {0}", _routingInfo.ServerName);
+                        SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.LoginWithFailover> Routed to {0}", RoutingInfo.ServerName);
 
                         if (_parser != null)
                         {
                             _parser.Disconnect();
                         }
 
-                        _parser = new TdsParser(ConnectionOptions.MARS, ConnectionOptions.Asynchronous);
+                        _parser = new TdsParser(ConnectionOptions.MARS, connectionOptions.Asynchronous);
+
                         Debug.Assert(SniContext.Undefined == Parser._physicalStateObj.SniContext, $"SniContext should be Undefined; actual Value: {Parser._physicalStateObj.SniContext}");
 
-                        currentServerInfo = new ServerInfo(ConnectionOptions, _routingInfo, currentServerInfo.ResolvedServerName, currentServerInfo.ServerSPN);
+                        currentServerInfo = new ServerInfo(ConnectionOptions, RoutingInfo, currentServerInfo.ResolvedServerName, currentServerInfo.ServerSPN);
                         _timeoutErrorInternal.SetInternalSourceType(SqlConnectionInternalSourceType.RoutingDestination);
                         _originalClientConnectionId = _clientConnectionId;
                         _routingDestination = currentServerInfo.UserServerName;
 
                         // restore properties that could be changed by the environment tokens
-                        _currentPacketSize = ConnectionOptions.PacketSize;
+                        _currentPacketSize = connectionOptions.PacketSize;
                         _currentLanguage = _originalLanguage = ConnectionOptions.CurrentLanguage;
-                        CurrentDatabase = _originalDatabase = ConnectionOptions.InitialCatalog;
+                        CurrentDatabase = _originalDatabase = connectionOptions.InitialCatalog;
                         _currentFailoverPartner = null;
                         _instanceName = string.Empty;
 
@@ -1999,8 +1983,7 @@ namespace Microsoft.Data.SqlClient
                                 newPassword,
                                 newSecurePassword,
                                 intervalTimer,
-                                withFailover: true
-                                );
+                                withFailover: true);
                     }
 
                     break; // leave the while loop -- we've successfully connected
@@ -2032,8 +2015,6 @@ namespace Microsoft.Data.SqlClient
                             throw;
                         }
                     }
-
-                    // TODO: Stash parser away somewhere so we can examine it's state during debugging
                 }
 
                 // We only get here when we failed to connect, but are going to re-try
@@ -2115,11 +2096,16 @@ namespace Microsoft.Data.SqlClient
         }
 
         // Common code path for making one attempt to establish a connection and log in to server.
-        private void AttemptOneLogin(ServerInfo serverInfo, string newPassword, SecureString newSecurePassword, TimeoutTimer timeout, bool withFailover = false, bool isFirstTransparentAttempt = true, bool disableTnir = false)
+        private void AttemptOneLogin(ServerInfo serverInfo,
+                                    string newPassword,
+                                    SecureString newSecurePassword,
+                                    TimeoutTimer timeout,
+                                    bool withFailover = false,
+                                    bool isFirstTransparentAttempt = true,
+                                    bool disableTnir = false)
         {
             SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.AttemptOneLogin|ADV> {0}, timeout={1}[msec], server={2}", ObjectID, timeout.MillisecondsRemaining, serverInfo.ExtendedServerName);
-
-            _routingInfo = null; // forget routing information
+            RoutingInfo = null; // forget routing information
 
             _parser._physicalStateObj.SniContext = SniContext.Snix_Connect;
 
@@ -2144,7 +2130,6 @@ namespace Microsoft.Data.SqlClient
 
             _timeoutErrorInternal.EndPhase(SqlConnectionTimeoutErrorPhase.PostLogin);
         }
-
 
         internal void FailoverPermissionDemand()
         {
@@ -2179,11 +2164,10 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        // called by SqlConnection.RepairConnection which is a relatevly expensive way of repair inner connection
+        // called by SqlConnection.RepairConnection which is a relatively expensive way of repair inner connection
         // prior to execution of request, used from EnlistTransaction, EnlistDistributedTransaction and ChangeDatabase
         internal bool GetSessionAndReconnectIfNeeded(SqlConnection parent, int timeout = 0)
         {
-
             Debug.Assert(!ThreadHasParserLockForClose, "Cannot call this method if caller has parser lock");
             if (ThreadHasParserLockForClose)
             {
@@ -2247,7 +2231,6 @@ namespace Microsoft.Data.SqlClient
             SqlConnection connection = Connection;
             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.BreakConnection|RES|CPOOL> {0}, Breaking connection.", ObjectID);
             DoomThisConnection();   // Mark connection as unusable, so it will be destroyed
-
             if (connection != null)
             {
                 connection.Close();
@@ -2258,7 +2241,7 @@ namespace Microsoft.Data.SqlClient
         { // true if we are only draining environment change tokens, used by TdsParser
             get
             {
-                return _routingInfo != null; // connection was routed, ignore rest of env change
+                return RoutingInfo != null; // connection was routed, ignore rest of env change
             }
         }
 
@@ -2268,7 +2251,7 @@ namespace Microsoft.Data.SqlClient
             switch (rec._type)
             {
                 case TdsEnums.ENV_DATABASE:
-                    // If connection is not open and recovery is not in progresss, store the server value as the original.
+                    // If connection is not open and recovery is not in progress, store the server value as the original.
                     if (!_fConnectionOpen && _recoverySessionData == null)
                     {
                         _originalDatabase = rec._newValue;
@@ -2278,13 +2261,13 @@ namespace Microsoft.Data.SqlClient
                     break;
 
                 case TdsEnums.ENV_LANG:
-                    // If connection is not open and recovery is not in progresss, store the server value as the original.
+                    // If connection is not open and recovery is not in progress, store the server value as the original.
                     if (!_fConnectionOpen && _recoverySessionData == null)
                     {
                         _originalLanguage = rec._newValue;
                     }
 
-                    _currentLanguage = rec._newValue; // TODO: finish this.
+                    _currentLanguage = rec._newValue;
                     break;
 
                 case TdsEnums.ENV_PACKETSIZE:
@@ -2353,7 +2336,7 @@ namespace Microsoft.Data.SqlClient
                     {
                         throw SQL.ROR_InvalidRoutingInfo(this);
                     }
-                    _routingInfo = rec._newRoutingInfo;
+                    RoutingInfo = rec._newRoutingInfo;
                     break;
 
                 default:
@@ -2365,7 +2348,6 @@ namespace Microsoft.Data.SqlClient
         internal void OnLoginAck(SqlLoginAck rec)
         {
             _loginAck = rec;
-            // UNDONE:  throw an error if this is not 7.0 or 7.1[5].
             if (_recoverySessionData != null)
             {
                 if (_recoverySessionData._tdsVersion != rec.tdsVersion)
@@ -2389,11 +2371,11 @@ namespace Microsoft.Data.SqlClient
                          || _credential != null
                          || _accessTokenCallback != null
                          || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryInteractive
+                         || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryDeviceCodeFlow
                          || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryManagedIdentity
                          || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryMSI
                          || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryDefault
                          || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryWorkloadIdentity
-                         || ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryDeviceCodeFlow
                          || (ConnectionOptions.Authentication == SqlAuthenticationMethod.ActiveDirectoryIntegrated && _fedAuthRequired),
                          "Credentials aren't provided for calling MSAL");
             Debug.Assert(fedAuthInfo != null, "info should not be null.");
@@ -2467,7 +2449,7 @@ namespace Microsoft.Data.SqlClient
                         Debug.Assert(!attemptRefreshTokenLocked || _fedAuthToken != null, "Either Lock should not have been obtained or _fedAuthToken should not be null.");
                         Debug.Assert(!attemptRefreshTokenLocked || _newDbConnectionPoolAuthenticationContext != null, "Either Lock should not have been obtained or _newDbConnectionPoolAuthenticationContext should not be null.");
 
-                        // Indicate in Bid Trace that we are successful with the update.
+                        // Indicate in EventSource Trace that we are successful with the update.
                         if (attemptRefreshTokenLocked)
                         {
                             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFedAuthInfo> {0}, The attempt to get a new access token succeeded under the locked mode.", ObjectID);
@@ -2578,7 +2560,6 @@ namespace Microsoft.Data.SqlClient
         /// <returns>SqlFedAuthToken</returns>
         internal SqlFedAuthToken GetFedAuthToken(SqlFedAuthInfo fedAuthInfo)
         {
-
             Debug.Assert(fedAuthInfo != null, "fedAuthInfo should not be null.");
 
             // No:of milliseconds to sleep for the inital back off.
@@ -2749,9 +2730,7 @@ namespace Microsoft.Data.SqlClient
                 // Deal with normal MsalExceptions.
                 catch (MsalException msalException)
                 {
-                    if (MsalError.UnknownError != msalException.ErrorCode
-                        || _timeout.IsExpired
-                        || _timeout.MillisecondsRemaining <= sleepInterval)
+                    if (MsalError.UnknownError != msalException.ErrorCode || _timeout.IsExpired || _timeout.MillisecondsRemaining <= sleepInterval)
                     {
                         SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.GetFedAuthToken.MSALException error:> {0}", msalException.ErrorCode);
 
@@ -2767,7 +2746,21 @@ namespace Microsoft.Data.SqlClient
                 // All other exceptions from MSAL/Azure Identity APIs
                 catch (Exception e)
                 {
-                    throw SqlException.CreateException(new() { new(0, (byte)0x00, (byte)TdsEnums.FATAL_ERROR_CLASS, ConnectionOptions.DataSource, e.Message, ActiveDirectoryAuthentication.MSALGetAccessTokenFunctionName, 0) }, "", this, e);
+                    throw SqlException.CreateException(
+                        new()
+                        {
+                            new(
+                                0,
+                                (byte)0x00,
+                                (byte)TdsEnums.FATAL_ERROR_CLASS,
+                                ConnectionOptions.DataSource,
+                                e.Message,
+                                ActiveDirectoryAuthentication.MSALGetAccessTokenFunctionName,
+                                0)
+                        },
+                        "",
+                        this,
+                        e);
                 }
             }
 
@@ -2786,12 +2779,9 @@ namespace Microsoft.Data.SqlClient
 
         internal void OnFeatureExtAck(int featureId, byte[] data)
         {
-            if (_routingInfo != null)
+            if (RoutingInfo != null && featureId != TdsEnums.FEATUREEXT_SQLDNSCACHING)
             {
-                if (TdsEnums.FEATUREEXT_SQLDNSCACHING != featureId)
-                {
-                    return;
-                }
+                return;
             }
 
             switch (featureId)
@@ -2845,6 +2835,25 @@ namespace Microsoft.Data.SqlClient
                         }
                         break;
                     }
+
+                case TdsEnums.FEATUREEXT_GLOBALTRANSACTIONS:
+                    {
+                        SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for GlobalTransactions", ObjectID);
+
+                        if (data.Length < 1)
+                        {
+                            SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Unknown version number for GlobalTransactions", ObjectID);
+                            throw SQL.ParsingError(ParsingErrorState.CorruptedTdsStream);
+                        }
+
+                        IsGlobalTransaction = true;
+                        if (1 == data[0])
+                        {
+                            IsGlobalTransactionsEnabledForServer = true;
+                        }
+                        break;
+                    }
+
                 case TdsEnums.FEATUREEXT_FEDAUTH:
                     {
                         SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for federated authentication", ObjectID);
@@ -2899,7 +2908,6 @@ namespace Microsoft.Data.SqlClient
                             }
 #endif
                         }
-
                         break;
                     }
                 case TdsEnums.FEATUREEXT_TCE:
@@ -2926,26 +2934,8 @@ namespace Microsoft.Data.SqlClient
 
                         if (data.Length > 1)
                         {
+                            // Extract the type of enclave being used by the server.
                             _parser.EnclaveType = Encoding.Unicode.GetString(data, 2, (data.Length - 2));
-                        }
-
-                        break;
-                    }
-
-                case TdsEnums.FEATUREEXT_GLOBALTRANSACTIONS:
-                    {
-                        SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for GlobalTransactions", ObjectID);
-
-                        if (data.Length < 1)
-                        {
-                            SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Unknown version number for GlobalTransactions", ObjectID);
-                            throw SQL.ParsingError(ParsingErrorState.CorruptedTdsStream);
-                        }
-
-                        IsGlobalTransaction = true;
-                        if (1 == data[0])
-                        {
-                            IsGlobalTransactionsEnabledForServer = true;
                         }
                         break;
                     }
@@ -2976,7 +2966,6 @@ namespace Microsoft.Data.SqlClient
                         if (data.Length < 1)
                         {
                             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Unknown token for DATACLASSIFICATION", ObjectID);
-
                             throw SQL.ParsingError(ParsingErrorState.CorruptedTdsStream);
                         }
                         byte supportedDataClassificationVersion = data[0];
@@ -3023,7 +3012,7 @@ namespace Microsoft.Data.SqlClient
                             IsSQLDNSCachingSupported = true;
                             _cleanSQLDNSCaching = false;
 
-                            if (_routingInfo != null)
+                            if (RoutingInfo != null)
                             {
                                 IsDNSCachingBeforeRedirectSupported = true;
                             }
@@ -3035,12 +3024,10 @@ namespace Microsoft.Data.SqlClient
                             _cleanSQLDNSCaching = true;
                         }
 
-                        // need to add more steps for phrase 2
+                        // need to add more steps for phase 2
                         // get IPv4 + IPv6 + Port number
                         // not put them in the DNS cache at this point but need to store them somewhere
-
                         // generate pendingSQLDNSObject and turn on IsSQLDNSRetryEnabled flag
-
                         break;
                     }
 
@@ -3121,14 +3108,14 @@ namespace Microsoft.Data.SqlClient
         {
             get
             {
-                return m_userServerName;
+                return _userServerName;
             }
             private set
             {
-                m_userServerName = value;
+                _userServerName = value;
             }
         }
-        private string m_userServerName;
+        private string _userServerName;
 
         internal readonly string PreRoutingServerName;
 
