@@ -9,9 +9,9 @@ using System.Diagnostics;
 using System.Threading;
 using System.Transactions;
 using Microsoft.Data.Common;
+using Microsoft.Data.Common.ConnectionString;
 using Microsoft.Data.ProviderBase;
 using Microsoft.Data.SqlClient.ConnectionPool;
-
 
 namespace Microsoft.Data.SqlClient
 {
