@@ -9,13 +9,13 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Net.Http.Headers;
-using System.Runtime.CompilerServices;
 using System.Security;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Transactions;
 using Microsoft.Data.Common;
+using Microsoft.Data.Common.ConnectionString;
 using Microsoft.Data.ProviderBase;
 using Microsoft.Data.SqlClient.ConnectionPool;
 using Microsoft.Identity.Client;
@@ -106,13 +106,12 @@ namespace Microsoft.Data.SqlClient
     {
         // https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/wiki/retry-after#simple-retry-for-errors-with-http-error-codes-500-600
         internal const int MsalHttpRetryStatusCode = 429;
+        // Connection re-route limit
+        internal const int MaxNumberOfRedirectRoute = 10;
 
         // CONNECTION AND STATE VARIABLES
         private readonly SqlConnectionPoolGroupProviderInfo _poolGroupProviderInfo; // will only be null when called for ChangePassword, or creating SSE User Instance
         private TdsParser _parser;
-
-        // Connection re-route limit
-        internal const int _maxNumberOfRedirectRoute = 10;
 
         private SqlLoginAck _loginAck;
         private SqlCredential _credential;
@@ -186,7 +185,7 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        private bool _DNSCachingBeforeRedirect = false;
+        private bool _dnsCachingBeforeRedirect = false;
 
         /// <summary>
         /// Get or set if the control ring send redirect token and feature ext ack with true for DNSCaching
@@ -195,11 +194,11 @@ namespace Microsoft.Data.SqlClient
         {
             get
             {
-                return _DNSCachingBeforeRedirect;
+                return _dnsCachingBeforeRedirect;
             }
             set
             {
-                _DNSCachingBeforeRedirect = value;
+                _dnsCachingBeforeRedirect = value;
             }
         }
 
@@ -212,7 +211,7 @@ namespace Microsoft.Data.SqlClient
         internal byte _tceVersionSupported;
 
         // The pool that this connection is associated with, if at all it is.
-        private DbConnectionPool _dbConnectionPool;
+        private IDbConnectionPool _dbConnectionPool;
 
         // This is used to preserve the authentication context object if we decide to cache it for subsequent connections in the same pool.
         // This will finally end up in _dbConnectionPool.AuthenticationContexts, but only after 1 successful login to SQL Server using this context.
@@ -432,8 +431,9 @@ namespace Microsoft.Data.SqlClient
         }
 
         // OTHER STATE VARIABLES AND REFERENCES
-
         internal Guid _clientConnectionId = Guid.Empty;
+
+        // Routing information (ROR)
         private Guid _originalClientConnectionId = Guid.Empty;
         private string _routingDestination = null;
         private readonly TimeoutTimer _timeout;
@@ -453,16 +453,35 @@ namespace Microsoft.Data.SqlClient
             SessionData reconnectSessionData = null,
             bool applyTransientFaultHandling = false,
             string accessToken = null,
-            DbConnectionPool pool = null,
+            IDbConnectionPool pool = null,
             Func<SqlAuthenticationParameters, CancellationToken,
             Task<SqlAuthenticationToken>> accessTokenCallback = null) : base(connectionOptions)
-
         {
 #if DEBUG
             if (reconnectSessionData != null)
             {
                 reconnectSessionData._debugReconnectDataApplied = true;
             }
+#if NETFRAMEWORK
+            try
+            {
+                // use this to help validate this object is only created after the following permission has been previously demanded in the current codepath
+                if (userConnectionOptions != null)
+                {
+                    // As mentioned above, userConnectionOptions may be different to connectionOptions, so we need to demand on the correct connection string
+                    userConnectionOptions.DemandPermission();
+                }
+                else
+                {
+                    connectionOptions.DemandPermission();
+                }
+            }
+            catch (System.Security.SecurityException)
+            {
+                System.Diagnostics.Debug.Assert(false, "unexpected SecurityException for current codepath");
+                throw;
+            }
+#endif
 #endif
             Debug.Assert(reconnectSessionData == null || connectionOptions.ConnectRetryCount > 0, "Reconnect data supplied with CR turned off");
 
@@ -512,6 +531,9 @@ namespace Microsoft.Data.SqlClient
             _parserLock.Wait(canReleaseFromAnyThread: false);
             ThreadHasParserLockForClose = true;   // In case of error, let ourselves know that we already own the parser lock
 
+#if NETFRAMEWORK
+            RuntimeHelpers.PrepareConstrainedRegions();
+#endif
             try
             {
                 _timeout = TimeoutTimer.StartSecondsTimeout(connectionOptions.ConnectTimeout);
@@ -524,7 +546,6 @@ namespace Microsoft.Data.SqlClient
                     try
                     {
                         OpenLoginEnlist(_timeout, connectionOptions, credential, newPassword, newSecurePassword, redirectedUserInstance);
-
                         break;
                     }
                     catch (SqlException sqlex)
@@ -543,6 +564,21 @@ namespace Microsoft.Data.SqlClient
                         }
                     }
                 }
+            }
+            catch (System.OutOfMemoryException)
+            {
+                DoomThisConnection();
+                throw;
+            }
+            catch (System.StackOverflowException)
+            {
+                DoomThisConnection();
+                throw;
+            }
+            catch (System.Threading.ThreadAbortException)
+            {
+                DoomThisConnection();
+                throw;
             }
             finally
             {
@@ -697,6 +733,7 @@ namespace Microsoft.Data.SqlClient
         {
             get
             {
+                // TODO: probably need to use a different method, but that's a different bug
                 bool result = FindLiveReader(null) == null; // can't prepare with a live data reader...
                 return result;
             }
@@ -719,6 +756,16 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
+        /// <summary>
+        /// Get boolean that specifies whether an enlisted transaction can be unbound from
+        /// the connection when that transaction completes.
+        /// </summary>
+        /// <value>
+        /// This override always returns false.
+        /// </value>
+        /// <remarks>
+        /// The SqlInternalConnectionTds.CheckEnlistedTransactionBinding method handles implicit unbinding for disposed transactions.
+        /// </remarks>
         protected override bool UnbindOnTransactionCompletion
         {
             get
@@ -738,7 +785,7 @@ namespace Microsoft.Data.SqlClient
         [SuppressMessage("Microsoft.Globalization", "CA1303:DoNotPassLiteralsAsLocalizedParameters")] // copied from Triaged.cs
         protected override void ChangeDatabaseInternal(string database)
         {
-            // Add brackets around database
+            // MDAC 73598 - add brackets around database
             database = SqlConnection.FixupDatabaseTransactionName(database);
             Task executeTask = _parser.TdsExecuteSQLBatch("use " + database, ConnectionOptions.ConnectTimeout, null, _parser._physicalStateObj, sync: true);
             Debug.Assert(executeTask == null, "Shouldn't get a task when doing sync writes");
@@ -798,7 +845,7 @@ namespace Microsoft.Data.SqlClient
                 {
                     // if MARS is on, then a datareader associated with the command exists
                     // or if MARS is off, then a datareader exists
-                    throw ADP.OpenReaderExists(parser.MARSOn);
+                    throw ADP.OpenReaderExists(parser.MARSOn); // MDAC 66411
                 }
                 else if (!parser.MARSOn && parser._physicalStateObj.HasPendingData)
                 {
@@ -869,12 +916,8 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
-        internal override bool IsConnectionAlive(bool throwOnException)
-        {
-            bool isAlive = false;
-            isAlive = _parser._physicalStateObj.IsConnectionAlive(throwOnException);
-            return isAlive;
-        }
+        internal override bool IsConnectionAlive(bool throwOnException) =>
+            _parser._physicalStateObj.IsConnectionAlive(throwOnException);
 
         ////////////////////////////////////////////////////////////////////////////////////////
         // POOLING METHODS
@@ -882,6 +925,10 @@ namespace Microsoft.Data.SqlClient
 
         protected override void Activate(Transaction transaction)
         {
+#if NETFRAMEWORK
+            FailoverPermissionDemand(); // Demand for unspecified failover pooled connections
+#endif
+
             // When we're required to automatically enlist in transactions and
             // there is one we enlist in it. On the other hand, if there isn't a
             // transaction and we are currently enlisted in one, then we
@@ -1023,8 +1070,7 @@ namespace Microsoft.Data.SqlClient
                     string transactionName,
                     System.Data.IsolationLevel iso,
                     SqlInternalTransaction internalTransaction,
-                    bool isDelegateControlRequest
-        )
+                    bool isDelegateControlRequest)
         {
             TdsEnums.TransactionManagerRequestType requestType = TdsEnums.TransactionManagerRequestType.Begin;
             TdsEnums.TransactionManagerIsolationLevel isoLevel = TdsEnums.TransactionManagerIsolationLevel.ReadCommitted;
@@ -1145,7 +1191,7 @@ namespace Microsoft.Data.SqlClient
                     }
                 }
 
-                //  _parser may be nulled out during TdsExecuteTransactionManagerRequest.
+                // SQLBU #406778 - _parser may be nulled out during TdsExecuteTransactionManagerRequest.
                 //  Only use local variable after this call.
                 _parser.TdsExecuteTransactionManagerRequest(null, requestType, transactionName, isoLevel,
                     ConnectionOptions.ConnectTimeout, internalTransaction, stateObj, isDelegateControlRequest);
@@ -1171,6 +1217,7 @@ namespace Microsoft.Data.SqlClient
 
         internal override void DelegatedTransactionEnded()
         {
+            // TODO: I don't like the way that this works, but I don't want to rototill the entire pooler to avoid this call.
             base.DelegatedTransactionEnded();
         }
 
@@ -1207,6 +1254,7 @@ namespace Microsoft.Data.SqlClient
                     SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.CompleteLogin|ERR> {0}, Server never sent the requested federated authentication info", ObjectID);
                     throw SQL.ParsingError(ParsingErrorState.FedAuthInfoNotReceived);
                 }
+
                 if (!_sessionRecoveryAcknowledged)
                 {
                     _currentSessionData = null;
@@ -1221,7 +1269,7 @@ namespace Microsoft.Data.SqlClient
                     _currentSessionData._initialCollation = _currentSessionData._collation;
                     _currentSessionData._initialLanguage = _currentLanguage;
                 }
-                bool isEncrypted = _parser.EncryptionOptions == EncryptionOptions.ON;
+                bool isEncrypted = (_parser.EncryptionOptions & EncryptionOptions.OPTIONS_MASK) == EncryptionOptions.ON;
                 if (_recoverySessionData != null)
                 {
                     if (_recoverySessionData._encrypted != isEncrypted)
@@ -1374,9 +1422,10 @@ namespace Microsoft.Data.SqlClient
                 requestedFeatures |= TdsEnums.FeatureExtension.AzureSQLSupport;
             }
 
-            // The SQLDNSCaching feature is implicitly set
+            // The SQLDNSCaching and JSON features are implicitly set
             requestedFeatures |= TdsEnums.FeatureExtension.SQLDNSCaching;
             requestedFeatures |= TdsEnums.FeatureExtension.JsonSupport;
+
             _parser.TdsLogin(login, requestedFeatures, _recoverySessionData, _fedAuthFeatureExtensionData, encrypt);
         }
 
@@ -1515,23 +1564,29 @@ namespace Microsoft.Data.SqlClient
             Debug.Assert(object.ReferenceEquals(connectionOptions, this.ConnectionOptions), "ConnectionOptions argument and property must be the same"); // consider removing the argument
             int routingAttempts = 0;
             ServerInfo originalServerInfo = serverInfo; // serverInfo may end up pointing to new object due to routing, original object is used to set CurrentDatasource
+
             SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.LoginNoFailover|ADV> {0}, host={1}", ObjectID, serverInfo.UserServerName);
+
             int sleepInterval = 100;  //milliseconds to sleep (back off) between attempts.
 
             ResolveExtendedServerName(serverInfo, !redirectedUserInstance, connectionOptions);
 
             long timeoutUnitInterval = 0;
 
-            if (connectionOptions.MultiSubnetFailover)
+            bool isParallel = connectionOptions.MultiSubnetFailover;
+
+            if (isParallel)
             {
+                float failoverTimeoutStep = ADP.FailoverTimeoutStep;
+
                 // Determine unit interval
                 if (timeout.IsInfinite)
                 {
-                    timeoutUnitInterval = checked((long)(ADP.FailoverTimeoutStep * (1000L * ADP.DefaultConnectionTimeout)));
+                    timeoutUnitInterval = checked((long)(failoverTimeoutStep * (1000L * ADP.DefaultConnectionTimeout)));
                 }
                 else
                 {
-                    timeoutUnitInterval = checked((long)(ADP.FailoverTimeoutStep * timeout.MillisecondsRemaining));
+                    timeoutUnitInterval = checked((long)(failoverTimeoutStep * timeout.MillisecondsRemaining));
                 }
             }
             // Only three ways out of this loop:
@@ -1545,14 +1600,18 @@ namespace Microsoft.Data.SqlClient
             //  back into the parser for the error cases.
             int attemptNumber = 0;
             TimeoutTimer intervalTimer = null;
+            TimeoutTimer attemptOneLoginTimeout = timeout;
+
             while (true)
             {
-                if (connectionOptions.MultiSubnetFailover)
+                if (isParallel)
                 {
-                    attemptNumber++;
+                    int multiplier = ++attemptNumber;
+
                     // Set timeout for this attempt, but don't exceed original timer
-                    long nextTimeoutInterval = checked(timeoutUnitInterval * attemptNumber);
+                    long nextTimeoutInterval = checked(timeoutUnitInterval * multiplier);
                     long milliseconds = timeout.MillisecondsRemaining;
+
                     if (nextTimeoutInterval > milliseconds)
                     {
                         nextTimeoutInterval = milliseconds;
@@ -1570,10 +1629,15 @@ namespace Microsoft.Data.SqlClient
 
                 try
                 {
+                    if (isParallel)
+                    {
+                        attemptOneLoginTimeout = intervalTimer;
+                    }
+
                     AttemptOneLogin(serverInfo,
                                     newPassword,
                                     newSecurePassword,
-                                    connectionOptions.MultiSubnetFailover ? intervalTimer : timeout);
+                                    attemptOneLoginTimeout);
 
                     if (connectionOptions.MultiSubnetFailover && ServerProvidedFailOverPartner != null)
                     {
@@ -1584,9 +1648,9 @@ namespace Microsoft.Data.SqlClient
                     if (RoutingInfo != null)
                     {
                         SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.LoginNoFailover> Routed to {0}", serverInfo.ExtendedServerName);
-                        if (routingAttempts > _maxNumberOfRedirectRoute)
+                        if (routingAttempts > MaxNumberOfRedirectRoute)
                         {
-                            throw SQL.ROR_RecursiveRoutingNotSupported(this, _maxNumberOfRedirectRoute);
+                            throw SQL.ROR_RecursiveRoutingNotSupported(this, MaxNumberOfRedirectRoute);
                         }
 
                         if (timeout.IsExpired)
@@ -1758,6 +1822,7 @@ namespace Microsoft.Data.SqlClient
             }
 
             // Initialize loop variables
+            bool failoverDemandDone = false; // have we demanded for partner information yet (as necessary)?
             int attemptNumber = 0;
 
             // Only three ways out of this loop:
@@ -1792,6 +1857,14 @@ namespace Microsoft.Data.SqlClient
                 ServerInfo currentServerInfo;
                 if (useFailoverHost)
                 {
+                    if (!failoverDemandDone)
+                    {
+#if NETFRAMEWORK
+                        FailoverPermissionDemand();
+#endif
+                        failoverDemandDone = true;
+                    }
+
                     // Primary server may give us a different failover partner than the connection string indicates.  Update it
                     if (ServerProvidedFailOverPartner != null && failoverServerInfo.ResolvedServerName != ServerProvidedFailOverPartner)
                     {
@@ -1817,18 +1890,19 @@ namespace Microsoft.Data.SqlClient
                             intervalTimer,
                             withFailover: true
                             );
-                    int routingAttemps = 0;
+
+                    int routingAttempts = 0;
                     while (RoutingInfo != null)
                     {
-                        if (routingAttemps > _maxNumberOfRedirectRoute)
+                        if (routingAttempts > MaxNumberOfRedirectRoute)
                         {
-                            throw SQL.ROR_RecursiveRoutingNotSupported(this, _maxNumberOfRedirectRoute);
+                            throw SQL.ROR_RecursiveRoutingNotSupported(this, MaxNumberOfRedirectRoute);
                         }
-                        routingAttemps++;
+                        routingAttempts++;
 
                         SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.LoginWithFailover> Routed to {0}", RoutingInfo.ServerName);
 
-                        if(_parser != null)
+                        if (_parser != null)
                         {
                             _parser.Disconnect();
                         }
@@ -1842,7 +1916,7 @@ namespace Microsoft.Data.SqlClient
                         _originalClientConnectionId = _clientConnectionId;
                         _routingDestination = currentServerInfo.UserServerName;
 
-                        // restore properties that could be changed by the environemnt tokens
+                        // restore properties that could be changed by the environment tokens
                         _currentPacketSize = connectionOptions.PacketSize;
                         _currentLanguage = _originalLanguage = ConnectionOptions.CurrentLanguage;
                         CurrentDatabase = _originalDatabase = connectionOptions.InitialCatalog;
@@ -1856,6 +1930,7 @@ namespace Microsoft.Data.SqlClient
                             intervalTimer,
                             withFailover: true);
                     }
+
                     break; // leave the while loop -- we've successfully connected
                 }
                 catch (SqlException sqlex)
@@ -1966,12 +2041,11 @@ namespace Microsoft.Data.SqlClient
         }
 
         // Common code path for making one attempt to establish a connection and log in to server.
-        private void AttemptOneLogin(
-                                ServerInfo serverInfo,
-                                string newPassword,
-                                SecureString newSecurePassword,
-                                TimeoutTimer timeout,
-                                bool withFailover = false)
+        private void AttemptOneLogin(ServerInfo serverInfo,
+                                    string newPassword,
+                                    SecureString newSecurePassword,
+                                    TimeoutTimer timeout,
+                                    bool withFailover = false)
         {
             SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.AttemptOneLogin|ADV> {0}, timeout={1}[msec], server={2}", ObjectID, timeout.MillisecondsRemaining, serverInfo.ExtendedServerName);
             RoutingInfo = null; // forget routing information 
@@ -1998,7 +2072,15 @@ namespace Microsoft.Data.SqlClient
             _timeoutErrorInternal.EndPhase(SqlConnectionTimeoutErrorPhase.PostLogin);
         }
 
-
+#if NETFRAMEWORK
+        internal void FailoverPermissionDemand()
+        {
+            if (PoolGroupProviderInfo != null)
+            {
+                PoolGroupProviderInfo.FailoverPermissionDemand();
+            }
+        }
+#endif
 
         ////////////////////////////////////////////////////////////////////////////////////////
         // PREPARED COMMAND METHODS
@@ -2041,18 +2123,39 @@ namespace Microsoft.Data.SqlClient
 
             try
             {
-                Task reconnectTask = parent.ValidateAndReconnect(() =>
+#if NETFRAMEWORK
+                RuntimeHelpers.PrepareConstrainedRegions();
+#endif
+                try
                 {
-                    ThreadHasParserLockForClose = false;
-                    _parserLock.Release();
-                    releaseConnectionLock = false;
-                }, timeout);
-                if (reconnectTask != null)
-                {
-                    AsyncHelper.WaitForCompletion(reconnectTask, timeout);
-                    return true;
+                    Task reconnectTask = parent.ValidateAndReconnect(() =>
+                    {
+                        ThreadHasParserLockForClose = false;
+                        _parserLock.Release();
+                        releaseConnectionLock = false;
+                    }, timeout);
+                    if (reconnectTask != null)
+                    {
+                        AsyncHelper.WaitForCompletion(reconnectTask, timeout);
+                        return true;
+                    }
+                    return false;
                 }
-                return false;
+                catch (System.OutOfMemoryException)
+                {
+                    DoomThisConnection();
+                    throw;
+                }
+                catch (System.StackOverflowException)
+                {
+                    DoomThisConnection();
+                    throw;
+                }
+                catch (System.Threading.ThreadAbortException)
+                {
+                    DoomThisConnection();
+                    throw;
+                }
             }
             finally
             {
@@ -2070,7 +2173,7 @@ namespace Microsoft.Data.SqlClient
 
         internal void BreakConnection()
         {
-            var connection = Connection;
+            SqlConnection connection = Connection;
             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.BreakConnection|RES|CPOOL> {0}, Breaking connection.", ObjectID);
             DoomThisConnection();   // Mark connection as unusable, so it will be destroyed
             if (connection != null)
@@ -2235,7 +2338,6 @@ namespace Microsoft.Data.SqlClient
             // We want to refresh the token, if taking the lock on the authentication context is successful.
             bool attemptRefreshTokenLocked = false;
 
-
             if (_dbConnectionPool != null)
             {
                 Debug.Assert(_dbConnectionPool.AuthenticationContexts != null);
@@ -2291,9 +2393,9 @@ namespace Microsoft.Data.SqlClient
                         // If the lock could not be obtained, it will return false, without attempting to fetch a new token.
                         attemptRefreshTokenLocked = TryGetFedAuthTokenLocked(fedAuthInfo, dbConnectionPoolAuthenticationContext, out _fedAuthToken);
 
-                        // If TryGetFedAuthTokenLocked returns true, it means lock was obtained and fedAuthToken should not be null.
+                        // If TryGetFedAuthTokenLocked returns true, it means lock was obtained and _fedAuthToken should not be null.
                         // If there was an exception in retrieving the new token, TryGetFedAuthTokenLocked should have thrown, so we won't be here.
-                        Debug.Assert(!attemptRefreshTokenLocked || _fedAuthToken != null, "Either Lock should not have been obtained or fedAuthToken should not be null.");
+                        Debug.Assert(!attemptRefreshTokenLocked || _fedAuthToken != null, "Either Lock should not have been obtained or _fedAuthToken should not be null.");
                         Debug.Assert(!attemptRefreshTokenLocked || _newDbConnectionPoolAuthenticationContext != null, "Either Lock should not have been obtained or _newDbConnectionPoolAuthenticationContext should not be null.");
 
                         // Indicate in EventSource Trace that we are successful with the update.
@@ -2311,7 +2413,7 @@ namespace Microsoft.Data.SqlClient
             {
                 // Get the Federated Authentication Token.
                 _fedAuthToken = GetFedAuthToken(fedAuthInfo);
-                Debug.Assert(_fedAuthToken != null, "fedAuthToken should not be null.");
+                Debug.Assert(_fedAuthToken != null, "_fedAuthToken should not be null.");
 
                 if (_dbConnectionPool != null)
                 {
@@ -2327,7 +2429,7 @@ namespace Microsoft.Data.SqlClient
             else if (!attemptRefreshTokenLocked)
             {
                 Debug.Assert(dbConnectionPoolAuthenticationContext != null, "dbConnectionPoolAuthenticationContext should not be null.");
-                Debug.Assert(_fedAuthToken == null, "fedAuthToken should be null in this case.");
+                Debug.Assert(_fedAuthToken == null, "_fedAuthToken should be null in this case.");
                 Debug.Assert(_newDbConnectionPoolAuthenticationContext == null, "_newDbConnectionPoolAuthenticationContext should be null.");
 
                 _fedAuthToken = new SqlFedAuthToken();
@@ -2338,7 +2440,7 @@ namespace Microsoft.Data.SqlClient
                 _fedAuthToken.expirationFileTime = dbConnectionPoolAuthenticationContext.ExpirationTime.ToFileTime();
             }
 
-            Debug.Assert(_fedAuthToken != null && _fedAuthToken.accessToken != null, "fedAuthToken and fedAuthToken.accessToken cannot be null.");
+            Debug.Assert(_fedAuthToken != null && _fedAuthToken.accessToken != null, "_fedAuthToken and _fedAuthToken.accessToken cannot be null.");
             _parser.SendFedAuthToken(_fedAuthToken);
         }
 
@@ -2423,7 +2525,7 @@ namespace Microsoft.Data.SqlClient
             // Username to use in error messages.
             string username = null;
 
-            var authProvider = SqlAuthenticationProvider.GetProvider(ConnectionOptions.Authentication);
+            SqlAuthenticationProvider authProvider = SqlAuthenticationProvider.GetProvider(ConnectionOptions.Authentication);
             if (authProvider == null && _accessTokenCallback == null)
                 throw SQL.CannotFindAuthProvider(ConnectionOptions.Authentication.ToString());
 
@@ -2597,14 +2699,8 @@ namespace Microsoft.Data.SqlClient
                         throw ADP.CreateSqlException(msalException, ConnectionOptions, this, username);
                     }
 
-                    SqlClientEventSource.Log.TryAdvancedTraceEvent(
-                        "<sc.SqlInternalConnectionTds.GetFedAuthToken|ADV> {0}, sleeping {1}[Milliseconds]",
-                        ObjectID,
-                        sleepInterval);
-                    SqlClientEventSource.Log.TryAdvancedTraceEvent(
-                        "<sc.SqlInternalConnectionTds.GetFedAuthToken|ADV> {0}, remaining {1}[Milliseconds]",
-                        ObjectID,
-                        _timeout.MillisecondsRemaining);
+                    SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.GetFedAuthToken|ADV> {0}, sleeping {1}[Milliseconds]", ObjectID, sleepInterval);
+                    SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.GetFedAuthToken|ADV> {0}, remaining {1}[Milliseconds]", ObjectID, _timeout.MillisecondsRemaining);
 
                     Thread.Sleep(sleepInterval);
                     sleepInterval *= 2;
@@ -2645,7 +2741,7 @@ namespace Microsoft.Data.SqlClient
 
         internal void OnFeatureExtAck(int featureId, byte[] data)
         {
-            if (RoutingInfo != null && TdsEnums.FEATUREEXT_SQLDNSCACHING != featureId)
+            if (RoutingInfo != null && featureId != TdsEnums.FEATUREEXT_SQLDNSCACHING)
             {
                 return;
             }
@@ -2705,6 +2801,7 @@ namespace Microsoft.Data.SqlClient
                 case TdsEnums.FEATUREEXT_GLOBALTRANSACTIONS:
                     {
                         SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for GlobalTransactions", ObjectID);
+
                         if (data.Length < 1)
                         {
                             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Unknown version number for GlobalTransactions", ObjectID);
@@ -2718,9 +2815,11 @@ namespace Microsoft.Data.SqlClient
                         }
                         break;
                     }
+
                 case TdsEnums.FEATUREEXT_FEDAUTH:
                     {
                         SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for federated authentication", ObjectID);
+
                         if (!_federatedAuthenticationRequested)
                         {
                             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Did not request federated authentication", ObjectID);
@@ -2743,6 +2842,7 @@ namespace Microsoft.Data.SqlClient
 
                             default:
                                 Debug.Fail("Unknown _fedAuthLibrary type");
+
                                 SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Attempting to use unknown federated authentication library", ObjectID);
                                 throw SQL.ParsingErrorLibraryType(ParsingErrorState.FedAuthFeatureAckUnknownLibraryType, (int)_fedAuthFeatureExtensionData.libraryType);
                         }
@@ -2770,7 +2870,6 @@ namespace Microsoft.Data.SqlClient
                             }
 #endif
                         }
-
                         break;
                     }
                 case TdsEnums.FEATUREEXT_TCE:
@@ -2802,6 +2901,7 @@ namespace Microsoft.Data.SqlClient
                         }
                         break;
                     }
+
                 case TdsEnums.FEATUREEXT_AZURESQLSUPPORT:
                     {
                         SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for AzureSQLSupport", ObjectID);
@@ -2817,10 +2917,10 @@ namespace Microsoft.Data.SqlClient
                         if ((data[0] & 1) == 1 && SqlClientEventSource.Log.IsTraceEnabled())
                         {
                             SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, FailoverPartner enabled with Readonly intent for AzureSQL DB", ObjectID);
-
                         }
                         break;
                     }
+
                 case TdsEnums.FEATUREEXT_DATACLASSIFICATION:
                     {
                         SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for DATACLASSIFICATION", ObjectID);
@@ -2828,7 +2928,6 @@ namespace Microsoft.Data.SqlClient
                         if (data.Length < 1)
                         {
                             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Unknown token for DATACLASSIFICATION", ObjectID);
-
                             throw SQL.ParsingError(ParsingErrorState.CorruptedTdsStream);
                         }
                         byte supportedDataClassificationVersion = data[0];
@@ -2841,13 +2940,13 @@ namespace Microsoft.Data.SqlClient
                         if (data.Length != 2)
                         {
                             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Unknown token for DATACLASSIFICATION", ObjectID);
-
                             throw SQL.ParsingError(ParsingErrorState.CorruptedTdsStream);
                         }
                         byte enabled = data[1];
                         _parser.DataClassificationVersion = (enabled == 0) ? TdsEnums.DATA_CLASSIFICATION_NOT_ENABLED : supportedDataClassificationVersion;
                         break;
                     }
+
                 case TdsEnums.FEATUREEXT_UTF8SUPPORT:
                     {
                         SqlClientEventSource.Log.TryAdvancedTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ADV> {0}, Received feature extension acknowledgement for UTF8 support", ObjectID);
@@ -2855,7 +2954,6 @@ namespace Microsoft.Data.SqlClient
                         if (data.Length < 1)
                         {
                             SqlClientEventSource.Log.TryTraceEvent("<sc.SqlInternalConnectionTds.OnFeatureExtAck|ERR> {0}, Unknown value for UTF8 support", ObjectID);
-
                             throw SQL.ParsingError();
                         }
                         break;
@@ -2892,7 +2990,6 @@ namespace Microsoft.Data.SqlClient
                         // get IPv4 + IPv6 + Port number
                         // not put them in the DNS cache at this point but need to store them somewhere
                         // generate pendingSQLDNSObject and turn on IsSQLDNSRetryEnabled flag
-
                         break;
                     }
 
