@@ -5,7 +5,8 @@
 using System;
 using System.Diagnostics;
 using Interop.Windows.Sni;
-using Microsoft.Data.SqlClient.SNI;
+using Microsoft.Data.SqlClient.ManagedSni;
+using SniError = Microsoft.Data.SqlClient.ManagedSni.SniError;
 
 namespace Microsoft.Data.SqlClient
 {
@@ -68,7 +69,7 @@ namespace Microsoft.Data.SqlClient
 
             if (TdsParserStateObjectFactory.UseManagedSNI)
             {
-                SNIError sniError = SNIProxy.Instance.GetLastError();
+                SniError sniError = SniProxy.Instance.GetLastError();
                 details.sniErrorNumber = sniError.sniError;
                 details.errorMessage = sniError.errorMessage;
                 details.nativeError = sniError.nativeError;
@@ -79,7 +80,7 @@ namespace Microsoft.Data.SqlClient
             }
             else
             {
-                SniNativeWrapper.SniGetLastError(out SniError sniError);
+                SniNativeWrapper.SniGetLastError(out Interop.Windows.Sni.SniError sniError);
                 details.sniErrorNumber = sniError.sniError;
                 details.errorMessage = sniError.errorMessage;
                 details.nativeError = sniError.nativeError;
