@@ -7,20 +7,19 @@
 using System;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
-using System.Net.Security;
 using System.Runtime.CompilerServices;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Data.Common;
 using Microsoft.Data.ProviderBase;
 
-namespace Microsoft.Data.SqlClient.SNI
+// @TODO: If this is in the manages SNI namespace, it should be in the managed SNI folder
+namespace Microsoft.Data.SqlClient.ManagedSni
 {
     internal sealed class TdsParserStateObjectManaged : TdsParserStateObject
     {
-        private SNIMarsConnection? _marsConnection;
-        private SNIHandle? _sessionHandle;
+        private SniMarsConnection? _marsConnection;
+        private SniHandle? _sessionHandle;
 
         public TdsParserStateObjectManaged(TdsParser parser) : base(parser) { }
 
@@ -34,7 +33,7 @@ namespace Microsoft.Data.SqlClient.SNI
 
         protected override bool CheckPacket(PacketHandle packet, TaskCompletionSource<object> source)
         {
-            SNIPacket p = packet.ManagedPacket;
+            SniPacket p = packet.ManagedPacket;
             return p.IsInvalid || source != null;
         }
 
@@ -52,7 +51,7 @@ namespace Microsoft.Data.SqlClient.SNI
             }
         }
 
-        internal SNIMarsHandle CreateMarsSession(object callbackObject, bool async)
+        internal SniMarsHandle CreateMarsSession(object callbackObject, bool async)
         {
             SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.CreateMarsSession | Info | State Object Id {0}, Session Id {1}, Async = {2}", _objectID, _sessionHandle?.ConnectionId, async);
             if (_marsConnection is null)
@@ -94,7 +93,7 @@ namespace Microsoft.Data.SqlClient.SNI
             string hostNameInCertificate,
             string serverCertificateFilename)
         {
-            SNIHandle? sessionHandle = SNIProxy.CreateConnectionHandle(serverName, timeout, out instanceName, ref spns, serverSPN,
+            SniHandle? sessionHandle = SniProxy.CreateConnectionHandle(serverName, timeout, out instanceName, ref spns, serverSPN,
                 flushCache, async, parallel, isIntegratedSecurity, iPAddressPreference, cachedFQDN, ref pendingDNSInfo, tlsFirst,
                 hostNameInCertificate, serverCertificateFilename);
 
@@ -120,9 +119,9 @@ namespace Microsoft.Data.SqlClient.SNI
             // No-op
         }
 
-        internal void ReadAsyncCallback(SNIPacket packet, uint error)
+        internal void ReadAsyncCallback(SniPacket packet, uint error)
         {
-            SNIHandle? sessionHandle = _sessionHandle;
+            SniHandle? sessionHandle = _sessionHandle;
             if (sessionHandle is not null)
             {
                 ReadAsyncCallback(IntPtr.Zero, PacketHandle.FromManagedPacket(packet), error);
@@ -140,9 +139,9 @@ namespace Microsoft.Data.SqlClient.SNI
             }
         }
 
-        internal void WriteAsyncCallback(SNIPacket packet, uint sniError)
+        internal void WriteAsyncCallback(SniPacket packet, uint sniError)
         {
-            SNIHandle? sessionHandle = _sessionHandle;
+            SniHandle? sessionHandle = _sessionHandle;
             if (sessionHandle is not null)
             {
                 WriteAsyncCallback(IntPtr.Zero, PacketHandle.FromManagedPacket(packet), sniError);
@@ -167,7 +166,7 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override void Dispose()
         {
-            SNIHandle? sessionHandle = Interlocked.Exchange(ref _sessionHandle, null);
+            SniHandle? sessionHandle = Interlocked.Exchange(ref _sessionHandle, null);
             if (sessionHandle is not null)
             {
                 SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.Dispose | Info | State Object Id {0}, Session Id {1}, Disposing session Handle and counters.", _objectID, sessionHandle.ConnectionId);
@@ -204,7 +203,7 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override bool IsFailedHandle()
         {
-            SNIHandle? sessionHandle = _sessionHandle;
+            SniHandle? sessionHandle = _sessionHandle;
             if (sessionHandle is not null)
             {
                 return sessionHandle.Status != TdsEnums.SNI_SUCCESS;
@@ -215,9 +214,9 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override PacketHandle ReadSyncOverAsync(int timeoutRemaining, out uint error)
         {
-            SNIHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
+            SniHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
 
-            error = sessionHandle.Receive(out SNIPacket packet, timeoutRemaining);
+            error = sessionHandle.Receive(out SniPacket packet, timeoutRemaining);
 
             SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.ReadSyncOverAsync | Info | State Object Id {0}, Session Id {1}", _objectID, sessionHandle.ConnectionId);
 #if DEBUG
@@ -234,14 +233,14 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override void ReleasePacket(PacketHandle syncReadPacket)
         {
-            SNIPacket packet = syncReadPacket.ManagedPacket;
+            SniPacket packet = syncReadPacket.ManagedPacket;
             SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.ReleasePacket | Info | State Object Id {0}, Session Id {1}, Packet DataLeft {2}", _objectID, _sessionHandle?.ConnectionId, packet?.DataLeft);
 #if DEBUG
             SqlClientEventSource.Log.TryAdvancedTraceEvent("TdsParserStateObjectManaged.ReleasePacket | TRC | State Object Id {0}, Session Id {1}, Packet {2} will be released, Packet Owner Id {3}, Packet dataLeft {4}", _objectID, _sessionHandle?.ConnectionId, packet?._id, packet?._owner.ConnectionId, packet?.DataLeft);
 #endif
             if (packet is not null)
             {
-                SNIHandle? sessionHandle = _sessionHandle;
+                SniHandle? sessionHandle = _sessionHandle;
                 if (sessionHandle is not null)
                 {
                     sessionHandle.ReturnPacket(packet);
@@ -257,13 +256,13 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override uint CheckConnection()
         {
-            SNIHandle? handle = GetSessionSNIHandleHandleOrThrow();
+            SniHandle? handle = GetSessionSNIHandleHandleOrThrow();
             return handle is null ? TdsEnums.SNI_SUCCESS : handle.CheckConnection();
         }
 
         internal override PacketHandle ReadAsync(SessionHandle handle, out uint error)
         {
-            SNIPacket? packet = null;
+            SniPacket? packet = null;
             error = handle.ManagedHandle.ReceiveAsync(ref packet);
 
             SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.ReadAsync | Info | State Object Id {0}, Session Id {1}, Packet DataLeft {2}", _objectID, _sessionHandle?.ConnectionId, packet?.DataLeft);
@@ -286,8 +285,8 @@ namespace Microsoft.Data.SqlClient.SNI
         internal override uint WritePacket(PacketHandle packetHandle, bool sync)
         {
             uint result = TdsEnums.SNI_UNINITIALIZED;
-            SNIHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
-            SNIPacket? packet = packetHandle.ManagedPacket;
+            SniHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
+            SniPacket? packet = packetHandle.ManagedPacket;
 
             if (sync)
             {
@@ -318,8 +317,8 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override PacketHandle GetResetWritePacket(int dataSize)
         {
-            SNIHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
-            SNIPacket packet = sessionHandle.RentPacket(headerSize: sessionHandle.ReserveHeaderSize, dataSize: dataSize);
+            SniHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
+            SniPacket packet = sessionHandle.RentPacket(headerSize: sessionHandle.ReserveHeaderSize, dataSize: dataSize);
 #if DEBUG
             Debug.Assert(packet.IsActive, "packet is not active, a serious pooling error may have occurred");
 #endif
@@ -346,7 +345,7 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override uint DisableSsl()
         {
-            SNIHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
+            SniHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
             SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.DisableSsl | Info | Session Id {0}", sessionHandle.ConnectionId);
             sessionHandle.DisableSsl();
             return TdsEnums.SNI_SUCCESS;
@@ -354,8 +353,8 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override uint EnableMars(ref uint info)
         {
-            SNIHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
-            _marsConnection = new SNIMarsConnection(sessionHandle);
+            SniHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
+            _marsConnection = new SniMarsConnection(sessionHandle);
             SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.EnableMars | Info | State Object Id {0}, Session Id {1}", _objectID, sessionHandle.ConnectionId);
 
             if (_marsConnection.StartReceive() == TdsEnums.SNI_SUCCESS_IO_PENDING)
@@ -368,7 +367,7 @@ namespace Microsoft.Data.SqlClient.SNI
 
         internal override uint EnableSsl(ref uint info, bool tlsFirst, string serverCertificateFilename)
         {
-            SNIHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
+            SniHandle sessionHandle = GetSessionSNIHandleHandleOrThrow();
             try
             {
                 SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.EnableSsl | Info | Session Id {0}", sessionHandle.ConnectionId);
@@ -377,7 +376,7 @@ namespace Microsoft.Data.SqlClient.SNI
             catch (Exception e)
             {
                 SqlClientEventSource.Log.TryTraceEvent("TdsParserStateObjectManaged.EnableSsl | Err | Session Id {0}, SNI Handshake failed with exception: {1}", sessionHandle.ConnectionId, e.Message);
-                return SNICommon.ReportSNIError(SNIProviders.SSL_PROV, SNICommon.HandshakeFailureError, e);
+                return SniCommon.ReportSNIError(SniProviders.SSL_PROV, SniCommon.HandshakeFailureError, e);
             }
         }
 
@@ -393,9 +392,9 @@ namespace Microsoft.Data.SqlClient.SNI
             return 0;
         }
 
-        private SNIHandle GetSessionSNIHandleHandleOrThrow()
+        private SniHandle GetSessionSNIHandleHandleOrThrow()
         {
-            SNIHandle? sessionHandle = _sessionHandle;
+            SniHandle? sessionHandle = _sessionHandle;
             if (sessionHandle is null)
             {
                 ThrowClosedConnection();
@@ -407,7 +406,7 @@ namespace Microsoft.Data.SqlClient.SNI
         [MethodImpl(MethodImplOptions.NoInlining)] // this forces the exception throwing code not to be inlined for performance
         private void ThrowClosedConnection() => throw ADP.ClosedConnectionError();
 
-        internal override SSPIContextProvider CreateSSPIContextProvider()
-            => new NegotiateSSPIContextProvider();
+        internal override SspiContextProvider CreateSspiContextProvider()
+            => new NegotiateSspiContextProvider();
     }
 }
