@@ -2,19 +2,20 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
+using System.Data;
+using System.Data.Common;
+using System.Diagnostics;
+using System.Runtime.ConstrainedExecution;
+using System.Threading;
+using System.Transactions;
+using Microsoft.Data.Common;
+using Microsoft.Data.Common.ConnectionString;
+using Microsoft.Data.ProviderBase;
+using Microsoft.Data.SqlClient.ConnectionPool;
+
 namespace Microsoft.Data.SqlClient
 {
-    using System;
-    using System.Data;
-    using System.Data.Common;
-    using System.Diagnostics;
-    using System.Runtime.ConstrainedExecution;
-    using System.Threading;
-    using System.Transactions;
-    using Microsoft.Data.Common;
-    using Microsoft.Data.ProviderBase;
-    using Microsoft.Data.SqlClient.ConnectionPool;
-
     public sealed partial class SqlConnection : DbConnection
     {
         private static readonly DbConnectionFactory _connectionFactory = SqlConnectionFactory.SingletonInstance;
