@@ -10,6 +10,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Data.Common;
+using Microsoft.Data.Common.ConnectionString;
 using Microsoft.Data.SqlClient;
 using Microsoft.Data.SqlClient.ConnectionPool;
 
@@ -18,7 +19,7 @@ namespace Microsoft.Data.ProviderBase
     internal abstract class DbConnectionFactory
     {
         private Dictionary<DbConnectionPoolKey, DbConnectionPoolGroup> _connectionPoolGroups;
-        private readonly List<DbConnectionPool> _poolsToRelease;
+        private readonly List<IDbConnectionPool> _poolsToRelease;
         private readonly List<DbConnectionPoolGroup> _poolGroupsToRelease;
         private readonly Timer _pruningTimer;
 
@@ -34,33 +35,13 @@ namespace Microsoft.Data.ProviderBase
         private static Task<DbConnectionInternal>[] s_pendingOpenNonPooled = new Task<DbConnectionInternal>[Environment.ProcessorCount];
         private static Task<DbConnectionInternal> s_completedTask;
 
-#if NETFRAMEWORK
-        private readonly DbConnectionPoolCounters _performanceCounters;
-
-        protected DbConnectionFactory() : this(DbConnectionPoolCountersNoCounters.SingletonInstance) { }
-
-        protected DbConnectionFactory(DbConnectionPoolCounters performanceCounters)
-        {
-            _performanceCounters = performanceCounters;
-            _connectionPoolGroups = new Dictionary<DbConnectionPoolKey, DbConnectionPoolGroup>();
-            _poolsToRelease = new List<DbConnectionPool>();
-            _poolGroupsToRelease = new List<DbConnectionPoolGroup>();
-            _pruningTimer = CreatePruningTimer();
-        }
-
-        internal DbConnectionPoolCounters PerformanceCounters
-        {
-            get { return _performanceCounters; }
-        }
-#else
         protected DbConnectionFactory()
         {
             _connectionPoolGroups = new Dictionary<DbConnectionPoolKey, DbConnectionPoolGroup>();
-            _poolsToRelease = new List<DbConnectionPool>();
+            _poolsToRelease = new List<IDbConnectionPool>();
             _poolGroupsToRelease = new List<DbConnectionPoolGroup>();
             _pruningTimer = CreatePruningTimer();
         }
-#endif
 
         public abstract DbProviderFactory ProviderFactory
         {
@@ -135,19 +116,14 @@ namespace Microsoft.Data.ProviderBase
             DbConnectionInternal newConnection = CreateConnection(connectionOptions, poolKey, poolGroupProviderInfo, null, owningConnection, userOptions);
             if (newConnection != null)
             {
-#if NETFRAMEWORK
-                PerformanceCounters.HardConnectsPerSecond.Increment();
-                newConnection.MakeNonPooledObject(owningConnection, PerformanceCounters);
-#else
-                SqlClientEventSource.Log.HardConnectRequest();
+                SqlClientEventSource.Metrics.HardConnectRequest();
                 newConnection.MakeNonPooledObject(owningConnection);
-#endif
             }
             SqlClientEventSource.Log.TryTraceEvent("<prov.DbConnectionFactory.CreateNonPooledConnection|RES|CPOOL> {0}, Non-pooled database connection created.", ObjectID);
             return newConnection;
         }
 
-        internal DbConnectionInternal CreatePooledConnection(DbConnectionPool pool, DbConnection owningObject, DbConnectionOptions options, DbConnectionPoolKey poolKey, DbConnectionOptions userOptions)
+        internal DbConnectionInternal CreatePooledConnection(IDbConnectionPool pool, DbConnection owningObject, DbConnectionOptions options, DbConnectionPoolKey poolKey, DbConnectionOptions userOptions)
         {
             Debug.Assert(pool != null, "null pool?");
             DbConnectionPoolGroupProviderInfo poolGroupProviderInfo = pool.PoolGroup.ProviderInfo;
@@ -155,11 +131,8 @@ namespace Microsoft.Data.ProviderBase
 
             if (newConnection != null)
             {
-#if NETFRAMEWORK
-                PerformanceCounters.HardConnectsPerSecond.Increment();
-#else
-                SqlClientEventSource.Log.HardConnectRequest();
-#endif
+                SqlClientEventSource.Metrics.HardConnectRequest();
+
                 newConnection.MakePooledConnection(pool);
             }
             SqlClientEventSource.Log.TryTraceEvent("<prov.DbConnectionFactory.CreatePooledConnection|RES|CPOOL> {0}, Pooled database connection created.", ObjectID);
@@ -204,7 +177,7 @@ namespace Microsoft.Data.ProviderBase
             Debug.Assert(owningConnection != null, "null owningConnection?");
 
             DbConnectionPoolGroup poolGroup;
-            DbConnectionPool connectionPool;
+            IDbConnectionPool connectionPool;
             connection = null;
 
             //  Work around race condition with clearing the pool between GetConnectionPool obtaining pool 
@@ -289,11 +262,8 @@ namespace Microsoft.Data.ProviderBase
                     }
 
                     connection = CreateNonPooledConnection(owningConnection, poolGroup, userOptions);
-#if NETFRAMEWORK
-                    PerformanceCounters.NumberOfNonPooledConnections.Increment();
-#else
-                    SqlClientEventSource.Log.EnterNonPooledConnection();
-#endif
+
+                    SqlClientEventSource.Metrics.EnterNonPooledConnection();
                 }
                 else
                 {
@@ -397,16 +367,12 @@ namespace Microsoft.Data.ProviderBase
                 }
                 else
                 {
-#if NETFRAMEWORK
-                    PerformanceCounters.NumberOfNonPooledConnections.Increment();
-#else
-                    SqlClientEventSource.Log.EnterNonPooledConnection();
-#endif
+                    SqlClientEventSource.Metrics.EnterNonPooledConnection();
                 }
             }
         }
 
-        private DbConnectionPool GetConnectionPool(DbConnection owningObject, DbConnectionPoolGroup connectionPoolGroup)
+        private IDbConnectionPool GetConnectionPool(DbConnection owningObject, DbConnectionPoolGroup connectionPoolGroup)
         {
             // if poolgroup is disabled, it will be replaced with a new entry
 
@@ -437,7 +403,7 @@ namespace Microsoft.Data.ProviderBase
                 Debug.Assert(connectionPoolGroup != null, "null connectionPoolGroup?");
                 SetConnectionPoolGroup(owningObject, connectionPoolGroup);
             }
-            DbConnectionPool connectionPool = connectionPoolGroup.GetConnectionPool(this);
+            IDbConnectionPool connectionPool = connectionPoolGroup.GetConnectionPool(this);
             return connectionPool;
         }
 
@@ -510,11 +476,8 @@ namespace Microsoft.Data.ProviderBase
 
                         // lock prevents race condition with PruneConnectionPoolGroups
                         newConnectionPoolGroups.Add(key, newConnectionPoolGroup);
-#if NETFRAMEWORK
-                        PerformanceCounters.NumberOfActiveConnectionPoolGroups.Increment();
-#else
-                        SqlClientEventSource.Log.EnterActiveConnectionPoolGroup();
-#endif
+
+                        SqlClientEventSource.Metrics.EnterActiveConnectionPoolGroup();
                         connectionPoolGroup = newConnectionPoolGroup;
                         _connectionPoolGroups = newConnectionPoolGroups;
                     }
@@ -568,8 +531,8 @@ namespace Microsoft.Data.ProviderBase
             {
                 if (0 != _poolsToRelease.Count)
                 {
-                    DbConnectionPool[] poolsToRelease = _poolsToRelease.ToArray();
-                    foreach (DbConnectionPool pool in poolsToRelease)
+                    IDbConnectionPool[] poolsToRelease = _poolsToRelease.ToArray();
+                    foreach (IDbConnectionPool pool in poolsToRelease)
                     {
                         if (pool != null)
                         {
@@ -578,12 +541,9 @@ namespace Microsoft.Data.ProviderBase
                             if (0 == pool.Count)
                             {
                                 _poolsToRelease.Remove(pool);
-                                SqlClientEventSource.Log.TryAdvancedTraceEvent("<prov.DbConnectionFactory.PruneConnectionPoolGroups|RES|INFO|CPOOL> {0}, ReleasePool={1}", ObjectID, pool.ObjectId);
-#if NETFRAMEWORK
-                                PerformanceCounters.NumberOfInactiveConnectionPools.Decrement();
-#else
-                                SqlClientEventSource.Log.ExitInactiveConnectionPool();
-#endif
+                                SqlClientEventSource.Log.TryAdvancedTraceEvent("<prov.DbConnectionFactory.PruneConnectionPoolGroups|RES|INFO|CPOOL> {0}, ReleasePool={1}", ObjectID, pool.Id);
+
+                                SqlClientEventSource.Metrics.ExitInactiveConnectionPool();
                             }
                         }
                     }
@@ -608,11 +568,8 @@ namespace Microsoft.Data.ProviderBase
                             {
                                 _poolGroupsToRelease.Remove(poolGroup);
                                 SqlClientEventSource.Log.TryAdvancedTraceEvent("<prov.DbConnectionFactory.PruneConnectionPoolGroups|RES|INFO|CPOOL> {0}, ReleasePoolGroup={1}", ObjectID, poolGroup.ObjectID);
-#if NETFRAMEWORK
-                                PerformanceCounters.NumberOfInactiveConnectionPoolGroups.Decrement();
-#else
-                                SqlClientEventSource.Log.ExitInactiveConnectionPoolGroup();
-#endif
+
+                                SqlClientEventSource.Metrics.ExitInactiveConnectionPoolGroup();
                             }
                         }
                     }
@@ -639,9 +596,6 @@ namespace Microsoft.Data.ProviderBase
                         if (entry.Value.Prune())
                         {
                             // may add entries to _poolsToRelease
-#if NETFRAMEWORK
-                            PerformanceCounters.NumberOfActiveConnectionPoolGroups.Decrement();
-#endif
                             QueuePoolGroupForRelease(entry.Value);
                         }
                         else
@@ -654,7 +608,7 @@ namespace Microsoft.Data.ProviderBase
             }
         }
 
-        internal void QueuePoolForRelease(DbConnectionPool pool, bool clearing)
+        internal void QueuePoolForRelease(IDbConnectionPool pool, bool clearing)
         {
             // Queue the pool up for release -- we'll clear it out and dispose
             // of it as the last part of the pruning timer callback so we don't
@@ -674,12 +628,8 @@ namespace Microsoft.Data.ProviderBase
                 }
                 _poolsToRelease.Add(pool);
             }
-#if NETFRAMEWORK
-            PerformanceCounters.NumberOfInactiveConnectionPools.Increment();
-#else
-            SqlClientEventSource.Log.EnterInactiveConnectionPool();
-            SqlClientEventSource.Log.ExitActiveConnectionPool();
-#endif
+            SqlClientEventSource.Metrics.EnterInactiveConnectionPool();
+            SqlClientEventSource.Metrics.ExitActiveConnectionPool();
         }
 
         internal void QueuePoolGroupForRelease(DbConnectionPoolGroup poolGroup)
@@ -691,20 +641,17 @@ namespace Microsoft.Data.ProviderBase
             {
                 _poolGroupsToRelease.Add(poolGroup);
             }
-#if NETFRAMEWORK
-            PerformanceCounters.NumberOfInactiveConnectionPoolGroups.Increment();
-#else
-            SqlClientEventSource.Log.EnterInactiveConnectionPoolGroup();
-            SqlClientEventSource.Log.ExitActiveConnectionPoolGroup();
-#endif
+
+            SqlClientEventSource.Metrics.EnterInactiveConnectionPoolGroup();
+            SqlClientEventSource.Metrics.ExitActiveConnectionPoolGroup();
         }
 
-        virtual protected DbConnectionInternal CreateConnection(DbConnectionOptions options, DbConnectionPoolKey poolKey, object poolGroupProviderInfo, DbConnectionPool pool, DbConnection owningConnection, DbConnectionOptions userOptions)
+        virtual protected DbConnectionInternal CreateConnection(DbConnectionOptions options, DbConnectionPoolKey poolKey, object poolGroupProviderInfo, IDbConnectionPool pool, DbConnection owningConnection, DbConnectionOptions userOptions)
         {
             return CreateConnection(options, poolKey, poolGroupProviderInfo, pool, owningConnection);
         }
 
-        abstract protected DbConnectionInternal CreateConnection(DbConnectionOptions options, DbConnectionPoolKey poolKey, object poolGroupProviderInfo, DbConnectionPool pool, DbConnection owningConnection);
+        abstract protected DbConnectionInternal CreateConnection(DbConnectionOptions options, DbConnectionPoolKey poolKey, object poolGroupProviderInfo, IDbConnectionPool pool, DbConnection owningConnection);
 
         abstract protected DbConnectionOptions CreateConnectionOptions(string connectionString, DbConnectionOptions previous);
 
