@@ -26,8 +26,7 @@ namespace Microsoft.Data.ProviderBase
         private const string PopulationStringKey = "PopulationString";
         private const string MaximumVersionKey = "MaximumVersion";
         private const string MinimumVersionKey = "MinimumVersion";
-        private const string DataSourceProductVersionNormalizedKey = "DataSourceProductVersionNormalized";
-        private const string DataSourceProductVersionKey = "DataSourceProductVersion";
+        private const string RestrictionDefaultKey = "RestrictionDefault";
         private const string RestrictionNumberKey = "RestrictionNumber";
         private const string NumberOfRestrictionsKey = "NumberOfRestrictions";
         private const string RestrictionNameKey = "RestrictionName";
@@ -44,10 +43,10 @@ namespace Microsoft.Data.ProviderBase
             ADP.CheckArgumentNull(serverVersion, nameof(serverVersion));
             ADP.CheckArgumentNull(normalizedServerVersion, nameof(normalizedServerVersion));
 
-            LoadDataSetFromXml(xmlStream);
-
             _serverVersionString = serverVersion;
             _normalizedServerVersion = normalizedServerVersion;
+
+            LoadDataSetFromXml(xmlStream);
         }
 
         protected DataSet CollectionDataSet => _metaDataCollectionsDataSet;
@@ -317,27 +316,13 @@ namespace Microsoft.Data.ProviderBase
 
         }
 
-        private void FixUpVersion(DataTable dataSourceInfoTable)
+        private void FixUpDataSourceInformationRow(DataRow dataSourceInfoRow)
         {
-            Debug.Assert(dataSourceInfoTable.TableName == DbMetaDataCollectionNames.DataSourceInformation);
-            DataColumn versionColumn = dataSourceInfoTable.Columns[DataSourceProductVersionKey];
-            DataColumn normalizedVersionColumn = dataSourceInfoTable.Columns[DataSourceProductVersionNormalizedKey];
+            Debug.Assert(dataSourceInfoRow.Table.Columns.Contains(DbMetaDataColumnNames.DataSourceProductVersion));
+            Debug.Assert(dataSourceInfoRow.Table.Columns.Contains(DbMetaDataColumnNames.DataSourceProductVersionNormalized));
 
-            if ((versionColumn == null) || (normalizedVersionColumn == null))
-            {
-                throw ADP.MissingDataSourceInformationColumn();
-            }
-
-            if (dataSourceInfoTable.Rows.Count != 1)
-            {
-                throw ADP.IncorrectNumberOfDataSourceInformationRows();
-            }
-
-            DataRow dataSourceInfoRow = dataSourceInfoTable.Rows[0];
-
-            dataSourceInfoRow[versionColumn] = _serverVersionString;
-            dataSourceInfoRow[normalizedVersionColumn] = _normalizedServerVersion;
-            dataSourceInfoRow.AcceptChanges();
+            dataSourceInfoRow[DbMetaDataColumnNames.DataSourceProductVersion] = _serverVersionString;
+            dataSourceInfoRow[DbMetaDataColumnNames.DataSourceProductVersionNormalized] = _normalizedServerVersion;
         }
 
 
@@ -441,15 +426,6 @@ namespace Microsoft.Data.ProviderBase
 
 
                     requestedSchema = CloneAndFilterCollection(exactCollectionName, hiddenColumns);
-
-                    // TODO: Consider an alternate method that doesn't involve special casing -- perhaps _prepareCollection
-
-                    // for the data source information table we need to fix up the version columns at run time
-                    // since the version is determined at run time
-                    if (exactCollectionName == DbMetaDataCollectionNames.DataSourceInformation)
-                    {
-                        FixUpVersion(requestedSchema);
-                    }
                     break;
 
                 case SqlCommandKey:
@@ -502,17 +478,228 @@ namespace Microsoft.Data.ProviderBase
 
         private void LoadDataSetFromXml(Stream XmlStream)
         {
-            _metaDataCollectionsDataSet = new DataSet
+            DataSet metaDataCollectionsDataSet = new DataSet("NewDataSet")
             {
                 Locale = CultureInfo.InvariantCulture
             };
             XmlReaderSettings settings = new()
             {
-                XmlResolver = null
+                XmlResolver = null,
+                IgnoreComments = true,
+                IgnoreWhitespace = true
             };
             using XmlReader reader = XmlReader.Create(XmlStream, settings);
-            _metaDataCollectionsDataSet.ReadXml(reader);
+
+            // Build up the schema DataSet manually, then load data from XmlStream. The schema of the DataSet is defined at:
+            // * https://learn.microsoft.com/en-us/sql/connect/ado-net/common-schema-collections
+            // * https://learn.microsoft.com/en-us/sql/connect/ado-net/sql-server-schema-collections
+            // Building the schema manually is necessary because DataSet.ReadXml uses XML serialization. This is slow, and it
+            // increases the binary size of AOT assemblies by approximately 4MB.
+
+            bool readContainer = reader.Read();
+
+            // Skip past the XML declaration and the outer container element. This XML document is hardcoded;
+            // these checks will need to be adjusted if its structure changes.
+            Debug.Assert(readContainer);
+            Debug.Assert(reader.NodeType == XmlNodeType.XmlDeclaration);
+
+            readContainer = reader.Read();
+            Debug.Assert(readContainer);
+            Debug.Assert(reader.NodeType == XmlNodeType.Element);
+            Debug.Assert(reader.Name == "MetaData");
+            Debug.Assert(reader.Depth == 0);
+
+            // Iterate over each "table element" of the outer container element.
+            // LoadDataTable will read the child elements of each table element.
+            while (reader.Read()
+                && reader.Depth == 1)
+            {
+                DataTable dataTable = null;
+                Action<DataRow> rowFixup = null;
+
+                Debug.Assert(reader.NodeType == XmlNodeType.Element);
+
+                switch (reader.Name)
+                {
+                    case "MetaDataCollectionsTable":
+                        dataTable = CreateMetaDataCollectionsDataTable();
+                        break;
+                    case "RestrictionsTable":
+                        dataTable = CreateRestrictionsDataTable();
+                        break;
+                    case "DataSourceInformationTable":
+                        dataTable = CreateDataSourceInformationDataTable();
+                        rowFixup = FixUpDataSourceInformationRow;
+                        break;
+                    case "DataTypesTable":
+                        dataTable = CreateDataTypesDataTable();
+                        break;
+                    case "ReservedWordsTable":
+                        dataTable = CreateReservedWordsDataTable();
+                        break;
+                    default:
+                        Debug.Fail($"Unexpected table element name: {reader.Name}");
+                        break;
+                }
+
+                if (dataTable != null)
+                {
+                    LoadDataTable(reader, dataTable, rowFixup);
+
+                    metaDataCollectionsDataSet.Tables.Add(dataTable);
+                }
+            }
+
+            _metaDataCollectionsDataSet = metaDataCollectionsDataSet;
         }
+
+        private static void LoadDataTable(XmlReader reader, DataTable table, Action<DataRow> rowFixup)
+        {
+            int parentDepth = reader.Depth;
+
+            table.BeginLoadData();
+
+            // One outer loop per element, each loop reading every property of the row
+            while (reader.Read()
+                && reader.Depth == parentDepth + 1)
+            {
+                Debug.Assert(reader.NodeType == XmlNodeType.Element);
+                Debug.Assert(reader.Name == table.TableName);
+
+                int childDepth = reader.Depth;
+                DataRow row = table.NewRow();
+
+                // Read every child property. Hardcoded structure - start with the element name, advance to the text, then to the EndElement
+                while (reader.Read()
+                    && reader.Depth == childDepth + 1)
+                {
+                    DataColumn column;
+                    bool successfulRead;
+
+                    Debug.Assert(reader.NodeType == XmlNodeType.Element);
+
+                    column = table.Columns[reader.Name];
+                    Debug.Assert(column != null);
+
+                    successfulRead = reader.Read();
+                    Debug.Assert(successfulRead);
+                    Debug.Assert(reader.NodeType == XmlNodeType.Text);
+
+                    row[column] = reader.Value;
+
+                    successfulRead = reader.Read();
+                    Debug.Assert(successfulRead);
+                    Debug.Assert(reader.NodeType == XmlNodeType.EndElement);
+                }
+
+                rowFixup?.Invoke(row);
+
+                table.Rows.Add(row);
+
+                Debug.Assert(reader.NodeType == XmlNodeType.EndElement);
+            }
+
+            table.EndLoadData();
+            table.AcceptChanges();
+        }
+
+        private static DataTable CreateMetaDataCollectionsDataTable()
+            => new DataTable(DbMetaDataCollectionNames.MetaDataCollections)
+            {
+                Columns =
+                {
+                    new DataColumn(DbMetaDataColumnNames.CollectionName, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.NumberOfRestrictions, typeof(int)),
+                    new DataColumn(DbMetaDataColumnNames.NumberOfIdentifierParts, typeof(int)),
+                    new DataColumn(PopulationMechanismKey, typeof(string)),
+                    new DataColumn(PopulationStringKey, typeof(string)),
+                    new DataColumn(MinimumVersionKey, typeof(string)),
+                    new DataColumn(MaximumVersionKey, typeof(string))
+                }
+            };
+
+        private static DataTable CreateRestrictionsDataTable()
+            => new DataTable(DbMetaDataCollectionNames.Restrictions)
+            {
+                Columns =
+                {
+                    new DataColumn(DbMetaDataColumnNames.CollectionName, typeof(string)),
+                    new DataColumn(RestrictionNameKey, typeof(string)),
+                    new DataColumn(ParameterNameKey, typeof(string)),
+                    new DataColumn(RestrictionDefaultKey, typeof(string)),
+                    new DataColumn(RestrictionNumberKey, typeof(int)),
+                    new DataColumn(MinimumVersionKey, typeof(string)),
+                    new DataColumn(MaximumVersionKey, typeof(string))
+                }
+            };
+
+        private static DataTable CreateDataSourceInformationDataTable()
+            => new DataTable(DbMetaDataCollectionNames.DataSourceInformation)
+            {
+                Columns =
+                {
+                    new DataColumn(DbMetaDataColumnNames.CompositeIdentifierSeparatorPattern, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.DataSourceProductName, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.DataSourceProductVersion, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.DataSourceProductVersionNormalized, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.GroupByBehavior, typeof(GroupByBehavior)),
+                    new DataColumn(DbMetaDataColumnNames.IdentifierPattern, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.IdentifierCase, typeof(IdentifierCase)),
+                    new DataColumn(DbMetaDataColumnNames.OrderByColumnsInSelect, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.ParameterMarkerFormat, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.ParameterMarkerPattern, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.ParameterNameMaxLength, typeof(int)),
+                    new DataColumn(DbMetaDataColumnNames.ParameterNamePattern, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.QuotedIdentifierPattern, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.QuotedIdentifierCase, typeof(IdentifierCase)),
+                    new DataColumn(DbMetaDataColumnNames.StatementSeparatorPattern, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.StringLiteralPattern, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.SupportedJoinOperators, typeof(SupportedJoinOperators))
+                }
+            };
+
+        private static DataTable CreateDataTypesDataTable()
+            => new DataTable(DbMetaDataCollectionNames.DataTypes)
+            {
+                Columns =
+                {
+                    new DataColumn(DbMetaDataColumnNames.TypeName, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.ProviderDbType, typeof(int)),
+                    new DataColumn(DbMetaDataColumnNames.ColumnSize, typeof(long)),
+                    new DataColumn(DbMetaDataColumnNames.CreateFormat, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.CreateParameters, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.DataType, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.IsAutoIncrementable, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsBestMatch, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsCaseSensitive, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsFixedLength, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsFixedPrecisionScale, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsLong, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsNullable, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsSearchable, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsSearchableWithLike, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.IsUnsigned, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.MaximumScale, typeof(short)),
+                    new DataColumn(DbMetaDataColumnNames.MinimumScale, typeof(short)),
+                    new DataColumn(DbMetaDataColumnNames.IsConcurrencyType, typeof(bool)),
+                    new DataColumn(MaximumVersionKey, typeof(string)),
+                    new DataColumn(MinimumVersionKey, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.IsLiteralSupported, typeof(bool)),
+                    new DataColumn(DbMetaDataColumnNames.LiteralPrefix, typeof(string)),
+                    new DataColumn(DbMetaDataColumnNames.LiteralSuffix, typeof(string))
+                }
+            };
+
+        private static DataTable CreateReservedWordsDataTable()
+            => new DataTable(DbMetaDataCollectionNames.ReservedWords)
+            {
+                Columns =
+                {
+                    new DataColumn(DbMetaDataColumnNames.ReservedWord, typeof(string)),
+                    new DataColumn(MinimumVersionKey, typeof(string)),
+                    new DataColumn(MaximumVersionKey, typeof(string))
+                }
+            };
 
         protected virtual DataTable PrepareCollection(string collectionName, string[] restrictions, DbConnection connection)
         {
