@@ -2,13 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
-using Microsoft.Data.Common;
-using Microsoft.Data.ProviderBase;
+using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading;
+using Microsoft.Data.Common;
+using Microsoft.Data.Common.ConnectionString;
+using Microsoft.Data.ProviderBase;
 
 namespace Microsoft.Data.SqlClient.ConnectionPool
 {
@@ -34,7 +35,7 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
         private readonly DbConnectionOptions _connectionOptions;
         private readonly DbConnectionPoolKey _poolKey;
         private readonly DbConnectionPoolGroupOptions _poolGroupOptions;
-        private ConcurrentDictionary<DbConnectionPoolIdentity, DbConnectionPool> _poolCollection;
+        private ConcurrentDictionary<DbConnectionPoolIdentity, IDbConnectionPool> _poolCollection;
 
         private int _state;          // see PoolGroupState* below
 
@@ -64,7 +65,7 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
             // HybridDictionary does not create any sub-objects until add
             // so it is safe to use for non-pooled connection as long as
             // we check _poolGroupOptions first
-            _poolCollection = new ConcurrentDictionary<DbConnectionPoolIdentity, DbConnectionPool>();
+            _poolCollection = new ConcurrentDictionary<DbConnectionPoolIdentity, IDbConnectionPool>();
             _state = PoolGroupStateActive;
         }
 
@@ -113,28 +114,26 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
             // will return the number of connections in the group after clearing has finished
 
             // First, note the old collection and create a new collection to be used
-            ConcurrentDictionary<DbConnectionPoolIdentity, DbConnectionPool> oldPoolCollection = null;
+            ConcurrentDictionary<DbConnectionPoolIdentity, IDbConnectionPool> oldPoolCollection = null;
             lock (this)
             {
                 if (_poolCollection.Count > 0)
                 {
                     oldPoolCollection = _poolCollection;
-                    _poolCollection = new ConcurrentDictionary<DbConnectionPoolIdentity, DbConnectionPool>();
+                    _poolCollection = new ConcurrentDictionary<DbConnectionPoolIdentity, IDbConnectionPool>();
                 }
             }
 
             // Then, if a new collection was created, release the pools from the old collection
             if (oldPoolCollection != null)
             {
-                foreach (KeyValuePair<DbConnectionPoolIdentity, DbConnectionPool> entry in oldPoolCollection)
+                foreach (KeyValuePair<DbConnectionPoolIdentity, IDbConnectionPool> entry in oldPoolCollection)
                 {
-                    DbConnectionPool pool = entry.Value;
+                    IDbConnectionPool pool = entry.Value;
                     if (pool != null)
                     {
                         DbConnectionFactory connectionFactory = pool.ConnectionFactory;
-#if NETFRAMEWORK
-                        connectionFactory.PerformanceCounters.NumberOfActiveConnectionPools.Decrement();
-#endif
+
                         connectionFactory.QueuePoolForRelease(pool, true);
                     }
                 }
@@ -144,7 +143,7 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
             return _poolCollection.Count;
         }
 
-        internal DbConnectionPool GetConnectionPool(DbConnectionFactory connectionFactory)
+        internal IDbConnectionPool GetConnectionPool(DbConnectionFactory connectionFactory)
         {
             // When this method returns null it indicates that the connection
             // factory should not use pooling.
@@ -152,7 +151,7 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
             // We don't support connection pooling on Win9x;
             // PoolGroupOptions will only be null when we're not supposed to pool
             // connections.
-            DbConnectionPool pool = null;
+            IDbConnectionPool pool = null;
             if (_poolGroupOptions != null)
             {
 #if NETFRAMEWORK
@@ -187,7 +186,17 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
                             if (!_poolCollection.TryGetValue(currentIdentity, out pool))
                             {
                                 DbConnectionPoolProviderInfo connectionPoolProviderInfo = connectionFactory.CreateConnectionPoolProviderInfo(ConnectionOptions);
-                                DbConnectionPool newPool = new WaitHandleDbConnectionPool(connectionFactory, this, currentIdentity, connectionPoolProviderInfo);
+
+                                IDbConnectionPool newPool;
+                                if (LocalAppContextSwitches.UseConnectionPoolV2)
+                                {
+                                    throw new NotImplementedException();
+                                }
+                                else
+                                {
+                                    // WaitHandleDbConnectionPool is the v1 pool implementation, and used by default if UseConnectionPoolV2 is off
+                                    newPool = new WaitHandleDbConnectionPool(connectionFactory, this, currentIdentity, connectionPoolProviderInfo);
+                                }
 
                                 if (MarkPoolGroupAsActive())
                                 {
@@ -197,10 +206,8 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
                                     newPool.Startup(); // must start pool before usage
                                     bool addResult = _poolCollection.TryAdd(currentIdentity, newPool);
                                     Debug.Assert(addResult, "No other pool with current identity should exist at this point");
-                                    SqlClientEventSource.Log.EnterActiveConnectionPool();
-#if NETFRAMEWORK
-                                    connectionFactory.PerformanceCounters.NumberOfActiveConnectionPools.Increment();
-#endif
+                                    SqlClientEventSource.Metrics.EnterActiveConnectionPool();
+
                                     pool = newPool;
                                 }
                                 else
@@ -257,11 +264,11 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
             {
                 if (_poolCollection.Count > 0)
                 {
-                    var newPoolCollection = new ConcurrentDictionary<DbConnectionPoolIdentity, DbConnectionPool>();
+                    var newPoolCollection = new ConcurrentDictionary<DbConnectionPoolIdentity, IDbConnectionPool>();
 
-                    foreach (KeyValuePair<DbConnectionPoolIdentity, DbConnectionPool> entry in _poolCollection)
+                    foreach (KeyValuePair<DbConnectionPoolIdentity, IDbConnectionPool> entry in _poolCollection)
                     {
-                        DbConnectionPool pool = entry.Value;
+                        IDbConnectionPool pool = entry.Value;
                         if (pool != null)
                         {
                             // Actually prune the pool if there are no connections in the pool and no errors occurred.
@@ -276,9 +283,7 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
                                 // pool into a list of pools to be released when they
                                 // are completely empty.
                                 DbConnectionFactory connectionFactory = pool.ConnectionFactory;
-#if NETFRAMEWORK
-                                connectionFactory.PerformanceCounters.NumberOfActiveConnectionPools.Decrement();
-#endif
+
                                 connectionFactory.QueuePoolForRelease(pool, false);
                             }
                             else
