@@ -12,11 +12,13 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Data.Common;
+#if NETFRAMEWORK
+using System.Runtime.ConstrainedExecution;
+#endif
 
 namespace Microsoft.Data.SqlClient
 {
 #if NETFRAMEWORK
-    using PacketHandle = IntPtr;
     using RuntimeHelpers = System.Runtime.CompilerServices.RuntimeHelpers;
 #endif
     
@@ -32,8 +34,10 @@ namespace Microsoft.Data.SqlClient
         InvalidData = 2
     }
 
-    partial class TdsParserStateObject
+    internal abstract partial class TdsParserStateObject
     {
+        private static readonly ContextCallback s_readAsyncCallbackComplete = ReadAsyncCallbackComplete;
+
         private static int s_objectTypeCount; // EventSource counter
         internal readonly int _objectID = Interlocked.Increment(ref s_objectTypeCount);
 
@@ -155,6 +159,7 @@ namespace Microsoft.Data.SqlClient
         internal volatile bool _attentionSent;              // true if we sent an Attention to the server
         internal volatile bool _attentionSending;
         private readonly TimerCallback _onTimeoutAsync;
+        private readonly WeakReference _cancellationOwner = new WeakReference(null);
 
         // Below 2 properties are used to enforce timeout delays in code to 
         // reproduce issues related to theadpool starvation and timeout delay.
@@ -304,7 +309,7 @@ namespace Microsoft.Data.SqlClient
         // Constructors //
         //////////////////
 
-        internal TdsParserStateObject(TdsParser parser)
+        protected TdsParserStateObject(TdsParser parser)
         {
             // Construct a physical connection
             Debug.Assert(parser != null, "no parser?");
@@ -338,6 +343,10 @@ namespace Microsoft.Data.SqlClient
             return (_snapshottedState & flag) == flag;
         }
 
+        ////////////////
+        // Properties //
+        ////////////////
+
         internal bool HasOpenResult
         {
             get => GetSnapshottedState(SnapshottedStateFlags.OpenResult);
@@ -367,10 +376,6 @@ namespace Microsoft.Data.SqlClient
             get => GetSnapshottedState(SnapshottedStateFlags.ColMetaDataReceived);
             set => SetSnapshottedState(SnapshottedStateFlags.ColMetaDataReceived, value);
         }
-
-        ////////////////
-        // Properties //
-        ////////////////
 
         internal int ObjectID => _objectID;
 
@@ -465,6 +470,26 @@ namespace Microsoft.Data.SqlClient
                 _timeoutTime = value;
             }
         }
+
+        ////////////////
+        // Properties //
+        ////////////////
+
+        internal abstract uint Status { get; }
+
+        internal abstract Guid? SessionId { get; }
+
+        internal abstract SessionHandle SessionHandle { get; }
+
+        internal abstract uint SniGetConnectionId(ref Guid clientConnectionId);
+
+        internal abstract uint DisableSsl();
+
+        internal abstract SspiContextProvider CreateSspiContextProvider();
+
+        internal abstract uint EnableMars(ref uint info);
+
+        internal abstract uint SetConnectionBufferSize(ref uint unsignedPacketSize);
 
         internal int GetTimeoutRemaining()
         {
@@ -676,6 +701,93 @@ namespace Microsoft.Data.SqlClient
             Debug.Assert(result == 1, "invalid deactivate count");
         }
 
+        // This method is only called by the command or datareader as a result of a user initiated
+        // cancel request.
+        internal void Cancel(object caller)
+        {
+            Debug.Assert(caller != null, "Null caller for Cancel!");
+            Debug.Assert(caller is SqlCommand || caller is SqlDataReader, "Calling API with invalid caller type: " + caller.GetType());
+
+            bool hasLock = false;
+            try
+            {
+                // Keep looping until we either grabbed the lock (and therefore sent attention) or the connection closes\breaks
+                while ((!hasLock) && (_parser.State != TdsParserState.Closed) && (_parser.State != TdsParserState.Broken))
+                {
+                    Monitor.TryEnter(this, WaitForCancellationLockPollTimeout, ref hasLock);
+                    if (hasLock)
+                    { // Lock for the time being - since we need to synchronize the attention send.
+                      // This lock is also protecting against concurrent close and async continuations
+
+                        // Ensure that, once we have the lock, that we are still the owner
+                        if ((!_cancelled) && (_cancellationOwner.Target == caller))
+                        {
+                            _cancelled = true;
+
+                            if (HasPendingData && !_attentionSent)
+                            {
+                                bool hasParserLock = false;
+                                // Keep looping until we have the parser lock (and so are allowed to write), or the connection closes\breaks
+                                while ((!hasParserLock) && (_parser.State != TdsParserState.Closed) && (_parser.State != TdsParserState.Broken))
+                                {
+                                    try
+                                    {
+                                        _parser.Connection._parserLock.Wait(canReleaseFromAnyThread: false, timeout: WaitForCancellationLockPollTimeout, lockTaken: ref hasParserLock);
+                                        if (hasParserLock)
+                                        {
+                                            _parser.Connection.ThreadHasParserLockForClose = true;
+                                            SendAttention();
+                                        }
+                                    }
+                                    finally
+                                    {
+                                        if (hasParserLock)
+                                        {
+                                            if (_parser.Connection.ThreadHasParserLockForClose)
+                                            {
+                                                _parser.Connection.ThreadHasParserLockForClose = false;
+                                            }
+                                            _parser.Connection._parserLock.Release();
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+            finally
+            {
+                if (hasLock)
+                {
+                    Monitor.Exit(this);
+                }
+            }
+        }
+
+        private void ResetCancelAndProcessAttention()
+        {
+            // This method is shared by CloseSession initiated by DataReader.Close or completed
+            // command execution, as well as the session reclamation code for cases where the
+            // DataReader is opened and then GC'ed.
+            lock (this)
+            {
+                // Reset cancel state.
+                _cancelled = false;
+                _cancellationOwner.Target = null;
+
+                if (_attentionSent)
+                {
+                    // Make sure we're cleaning up the AttentionAck if Cancel happened before taking the lock.
+                    // We serialize Cancel/CloseSession to prevent a race condition between these two states.
+                    // The problem is that both sending and receiving attentions are time taking
+                    // operations.
+                    Parser.ProcessPendingAck(this);
+                }
+                SetTimeoutStateStopped();
+            }
+        }
+
         // CancelRequest - use to cancel while writing a request to the server
         //
         // o none of the request might have been sent to the server, simply reset the buffer,
@@ -743,6 +855,11 @@ namespace Microsoft.Data.SqlClient
             }
         }
 
+        internal void StartSession(object cancellationOwner)
+        {
+            _cancellationOwner.Target = cancellationOwner;
+        }
+
         internal void CloseSession()
         {
             ResetCancelAndProcessAttention();
@@ -750,6 +867,17 @@ namespace Microsoft.Data.SqlClient
             InvalidateDebugOnlyCopyOfSniContext();
 #endif
             Parser.PutSession(this);
+        }
+
+#if NETFRAMEWORK
+        [ReliabilityContract(Consistency.WillNotCorruptState, Cer.Success)]
+#endif
+        internal int IncrementPendingCallbacks()
+        {
+            int remaining = Interlocked.Increment(ref _pendingCallbacks);
+            SqlClientEventSource.Log.TryAdvancedTraceEvent("TdsParserStateObject.IncrementPendingCallbacks | ADV | State Object Id {0}, after incrementing _pendingCallbacks: {1}", _objectID, _pendingCallbacks);
+            Debug.Assert(0 < remaining && remaining <= 3, $"_pendingCallbacks values is invalid after incrementing: {remaining}");
+            return remaining;
         }
 
         internal bool Deactivate()
@@ -876,6 +1004,65 @@ namespace Microsoft.Data.SqlClient
             {
                 _timeoutMilliseconds = timeout;
                 _timeoutTime = 0;
+            }
+        }
+
+        private void ChangeNetworkPacketTimeout(int dueTime, int period)
+        {
+            Timer networkPacketTimeout = _networkPacketTimeout;
+            if (networkPacketTimeout != null)
+            {
+                try
+                {
+                    networkPacketTimeout.Change(dueTime, period);
+                }
+                catch (ObjectDisposedException)
+                {
+                    // _networkPacketTimeout is set to null before Disposing, but there is still a slight chance
+                    // that object was disposed after we took a copy
+                }
+            }
+        }
+
+        private void ReadAsyncCallbackCaptureException(TaskCompletionSource<object> source)
+        {
+            bool captureSuccess = false;
+            try
+            {
+                if (_hasErrorOrWarning)
+                {
+                    // Do the close on another thread, since we don't want to block the callback thread
+                    ThrowExceptionAndWarning(asyncClose: true);
+                }
+                else if ((_parser.State == TdsParserState.Closed) || (_parser.State == TdsParserState.Broken))
+                {
+                    // Connection was closed by another thread before we parsed the packet, so no error was added to the collection
+                    throw ADP.ClosedConnectionError();
+                }
+            }
+            catch (Exception ex)
+            {
+                if (source.TrySetException(ex))
+                {
+                    // There was an exception, and it was successfully stored in the task
+                    captureSuccess = true;
+                }
+            }
+
+            if (!captureSuccess)
+            {
+                // Either there was no exception, or the task was already completed
+                // This is unusual, but possible if a fatal timeout occurred on another thread (which should mean that the connection is now broken)
+                Debug.Assert(_parser.State == TdsParserState.Broken || _parser.State == TdsParserState.Closed || _parser.Connection.IsConnectionDoomed, "Failed to capture exception while the connection was still healthy");
+
+                // The safest thing to do is to ensure that the connection is broken and attempt to cancel the task
+                // This must be done from another thread to not block the callback thread
+                Task.Factory.StartNew(() =>
+                {
+                    _parser.State = TdsParserState.Broken;
+                    _parser.Connection.BreakConnection();
+                    source.TrySetCanceled();
+                });
             }
         }
 
@@ -1683,7 +1870,7 @@ namespace Microsoft.Data.SqlClient
             }
             byte[] buf = null;
             int offset = 0;
-            (bool isAvailable, bool isStarting, bool isContinuing) = GetSnapshotStatuses();
+            (bool canContinue, bool isStarting, bool isContinuing) = GetSnapshotStatuses();
 
             if (isPlp)
             {
@@ -1701,7 +1888,7 @@ namespace Microsoft.Data.SqlClient
                 if (((_inBytesUsed + length) > _inBytesRead) || (_inBytesPacket < length))
                 {
                     int startOffset = 0;
-                    if (isAvailable)
+                    if (canContinue)
                     {
                         if (isContinuing || isStarting)
                         {
@@ -1719,7 +1906,7 @@ namespace Microsoft.Data.SqlClient
                         buf = new byte[length];
                     }
 
-                    TdsOperationStatus result = TryReadByteArray(buf, length, out _, startOffset, isAvailable);
+                    TdsOperationStatus result = TryReadByteArray(buf, length, out _, startOffset, canContinue);
                     
                     if (result != TdsOperationStatus.Done)
                     {
@@ -1849,14 +2036,15 @@ namespace Microsoft.Data.SqlClient
 
         internal TdsOperationStatus TryReadPlpBytes(ref byte[] buff, int offset, int len, out int totalBytesRead)
         {
+            bool canContinue = false;
             bool isStarting = false;
             bool isContinuing = false;
             bool compatibilityMode = LocalAppContextSwitches.UseCompatibilityAsyncBehaviour;
             if (!compatibilityMode)
             {
-                (_, isStarting, isContinuing) = GetSnapshotStatuses();
+                (canContinue, isStarting, isContinuing) = GetSnapshotStatuses();
             }
-            return TryReadPlpBytes(ref buff, offset, len, out totalBytesRead, isStarting || isContinuing, compatibilityMode);
+            return TryReadPlpBytes(ref buff, offset, len, out totalBytesRead, canContinue, canContinue, compatibilityMode);
         }
         // Reads the requested number of bytes from a plp data stream, or the entire data if
         // requested length is -1 or larger than the actual length of data. First call to this method
@@ -1864,7 +2052,7 @@ namespace Microsoft.Data.SqlClient
         // Returns the actual bytes read.
         // NOTE: This method must be retriable WITHOUT replaying a snapshot
         // Every time you call this method increment the offset and decrease len by the value of totalBytesRead
-        internal TdsOperationStatus TryReadPlpBytes(ref byte[] buff, int offset, int len, out int totalBytesRead, bool writeDataSizeToSnapshot, bool compatibilityMode)
+        internal TdsOperationStatus TryReadPlpBytes(ref byte[] buff, int offset, int len, out int totalBytesRead, bool canContinue, bool writeDataSizeToSnapshot, bool compatibilityMode)
         {
             totalBytesRead = 0;
 
@@ -1889,9 +2077,16 @@ namespace Microsoft.Data.SqlClient
             // If total length is known up front, allocate the whole buffer in one shot instead of realloc'ing and copying over each time
             if (buff == null && _longlen != TdsEnums.SQL_PLP_UNKNOWNLEN)
             {
-                if (writeDataSizeToSnapshot)
+                if (compatibilityMode && _snapshot != null && _snapshotStatus != SnapshotStatus.NotActive)
                 {
-                    // if there is a snapshot and it contains a stored plp buffer take it
+                    // legacy replay path perf optimization
+                    // if there is a snapshot which contains a stored plp buffer take it
+                    // and try to use it if it is the right length
+                    buff = TryTakeSnapshotStorage() as byte[];
+                }
+                else if (writeDataSizeToSnapshot && canContinue && _snapshot != null)
+                {
+                    // if there is a snapshot which it contains a stored plp buffer take it
                     // and try to use it if it is the right length
                     buff = TryTakeSnapshotStorage() as byte[];
                     if (buff != null)
@@ -1900,13 +2095,7 @@ namespace Microsoft.Data.SqlClient
                         totalBytesRead = offset;
                     }
                 }
-                else if (compatibilityMode && _snapshot != null && _snapshotStatus != SnapshotStatus.NotActive)
-                {
-                    // legacy replay path perf optimization
-                    // if there is a snapshot and it contains a stored plp buffer take it
-                    // and try to use it if it is the right length
-                    buff = TryTakeSnapshotStorage() as byte[];
-                }
+
 
                 if ((ulong)(buff?.Length ?? 0) != _longlen)
                 {
@@ -1958,22 +2147,28 @@ namespace Microsoft.Data.SqlClient
                 _longlenleft -= (ulong)bytesRead;
                 if (result != TdsOperationStatus.Done)
                 {
-                    if (writeDataSizeToSnapshot)
-                    {
-                        // a partial read has happened so store the target buffer in the snapshot
-                        // so it can be re-used when another packet arrives and we read again
-                        SetSnapshotStorage(buff);
-                        SetSnapshotDataSize(bytesRead);
-
-                    }
-                    else if (compatibilityMode && _snapshot != null)
+                    if (compatibilityMode && _snapshot != null)
                     {
                         // legacy replay path perf optimization
                         // a partial read has happened so store the target buffer in the snapshot
                         // so it can be re-used when another packet arrives and we read again
                         SetSnapshotStorage(buff);
                     }
+                    else if (canContinue)
+                    {
+                        // a partial read has happened so store the target buffer in the snapshot
+                        // so it can be re-used when another packet arrives and we read again
+                        SetSnapshotStorage(buff);
+                        if (writeDataSizeToSnapshot)
+                        {
+                            SetSnapshotDataSize(bytesRead);
+                        }
+                    }
                     return result;
+                }
+                if (writeDataSizeToSnapshot && canContinue)
+                {
+                    SetSnapshotDataSize(bytesRead);
                 }
 
                 if (_longlenleft == 0)
@@ -1982,19 +2177,19 @@ namespace Microsoft.Data.SqlClient
                     result = TryReadPlpLength(false, out _);
                     if (result != TdsOperationStatus.Done)
                     {
-                        if (writeDataSizeToSnapshot)
-                        {
-                            if (result == TdsOperationStatus.NeedMoreData)
-                            {
-                                SetSnapshotStorage(buff);
-                                SetSnapshotDataSize(bytesRead);
-                            }
-                        } 
-                        else if (compatibilityMode && _snapshot != null)
+                        if (compatibilityMode && _snapshot != null)
                         {
                             // a partial read has happened so store the target buffer in the snapshot
                             // so it can be re-used when another packet arrives and we read again
                             SetSnapshotStorage(buff);
+                        }
+                        else if (canContinue && result == TdsOperationStatus.NeedMoreData)
+                        {
+                            SetSnapshotStorage(buff);
+                            if (writeDataSizeToSnapshot)
+                            {
+                                SetSnapshotDataSize(bytesRead);
+                            }
                         }
                         return result;
                     }
@@ -2041,6 +2236,387 @@ namespace Microsoft.Data.SqlClient
         {
             Debug.Assert(_syncOverAsync || !_asyncReadWithoutSnapshot, "This method is not safe to call when doing sync over async");
             return TryReadByteArray(Span<byte>.Empty, num);
+        }
+
+        //////////////////////////////////////////////
+        // Statistics, Tracing, and related methods //
+        //////////////////////////////////////////////
+
+        private void SniReadStatisticsAndTracing()
+        {
+            SqlStatistics statistics = Parser.Statistics;
+            if (statistics != null)
+            {
+                if (statistics.WaitForReply)
+                {
+                    statistics.SafeIncrement(ref statistics._serverRoundtrips);
+                    statistics.ReleaseAndUpdateNetworkServerTimer();
+                }
+
+                statistics.SafeAdd(ref statistics._bytesReceived, _inBytesRead);
+                statistics.SafeIncrement(ref statistics._buffersReceived);
+            }
+        }
+
+        [Conditional("DEBUG")]
+        private void AssertValidState()
+        {
+            if (_inBytesUsed < 0 || _inBytesRead < 0)
+            {
+                Debug.Fail($"Invalid TDS Parser State: either _inBytesUsed or _inBytesRead is negative: {_inBytesUsed}, {_inBytesRead}");
+            }
+            else if (_inBytesUsed > _inBytesRead)
+            {
+                Debug.Fail($"Invalid TDS Parser State: _inBytesUsed > _inBytesRead: {_inBytesUsed} > {_inBytesRead}");
+            }
+
+            Debug.Assert(_inBytesPacket >= 0, "Packet must not be negative");
+        }
+
+        //////////////////////////////////////////////
+        // Errors and Warnings                      //
+        //////////////////////////////////////////////
+
+        /// <summary>
+        /// True if there is at least one error or warning (not counting the pre-attention errors\warnings)
+        /// </summary>
+        internal bool HasErrorOrWarning
+        {
+            get
+            {
+                return _hasErrorOrWarning;
+            }
+        }
+
+        /// <summary>
+        /// Adds an error to the error collection
+        /// </summary>
+        /// <param name="error"></param>
+        internal void AddError(SqlError error)
+        {
+            Debug.Assert(error != null, "Trying to add a null error");
+
+            // Switch to sync once we see an error
+            _syncOverAsync = true;
+
+            lock (_errorAndWarningsLock)
+            {
+                _hasErrorOrWarning = true;
+                if (_errors == null)
+                {
+                    _errors = new SqlErrorCollection();
+                }
+                _errors.Add(error);
+            }
+        }
+
+        /// <summary>
+        /// Gets the number of errors currently in the error collection
+        /// </summary>
+        internal int ErrorCount
+        {
+            get
+            {
+                int count = 0;
+                lock (_errorAndWarningsLock)
+                {
+                    if (_errors != null)
+                    {
+                        count = _errors.Count;
+                    }
+                }
+                return count;
+            }
+        }
+
+        /// <summary>
+        /// Adds an warning to the warning collection
+        /// </summary>
+        /// <param name="error"></param>
+        internal void AddWarning(SqlError error)
+        {
+            Debug.Assert(error != null, "Trying to add a null error");
+
+            // Switch to sync once we see a warning
+            _syncOverAsync = true;
+
+            lock (_errorAndWarningsLock)
+            {
+                _hasErrorOrWarning = true;
+                if (_warnings == null)
+                {
+                    _warnings = new SqlErrorCollection();
+                }
+                _warnings.Add(error);
+            }
+        }
+
+        /// <summary>
+        /// Gets the number of warnings currently in the warning collection
+        /// </summary>
+        internal int WarningCount
+        {
+            get
+            {
+                int count = 0;
+                lock (_errorAndWarningsLock)
+                {
+                    if (_warnings != null)
+                    {
+                        count = _warnings.Count;
+                    }
+                }
+                return count;
+            }
+        }
+
+        internal int PreAttentionErrorCount
+        {
+            get
+            {
+                int count = 0;
+                lock (_errorAndWarningsLock)
+                {
+                    if (_preAttentionErrors != null)
+                    {
+                        count = _preAttentionErrors.Count;
+                    }
+                }
+                return count;
+            }
+        }
+
+        /// <summary>
+        /// Gets the number of errors currently in the pre-attention warning collection
+        /// </summary>
+        internal int PreAttentionWarningCount
+        {
+            get
+            {
+                int count = 0;
+                lock (_errorAndWarningsLock)
+                {
+                    if (_preAttentionWarnings != null)
+                    {
+                        count = _preAttentionWarnings.Count;
+                    }
+                }
+                return count;
+            }
+        }
+
+        /// <summary>
+        /// Gets the full list of errors and warnings (including the pre-attention ones), then wipes all error and warning lists
+        /// </summary>
+        /// <param name="broken">If true, the connection should be broken</param>
+        /// <returns>An array containing all of the errors and warnings</returns>
+        internal SqlErrorCollection GetFullErrorAndWarningCollection(out bool broken)
+        {
+            SqlErrorCollection allErrors = new SqlErrorCollection();
+            broken = false;
+
+            lock (_errorAndWarningsLock)
+            {
+                _hasErrorOrWarning = false;
+
+                // Merge all error lists, then reset them
+                AddErrorsToCollection(_errors, ref allErrors, ref broken);
+                AddErrorsToCollection(_warnings, ref allErrors, ref broken);
+                _errors = null;
+                _warnings = null;
+
+                // We also process the pre-attention error lists here since, if we are here and they are populated, then an error occurred while sending attention so we should show the errors now (otherwise they'd be lost)
+                AddErrorsToCollection(_preAttentionErrors, ref allErrors, ref broken);
+                AddErrorsToCollection(_preAttentionWarnings, ref allErrors, ref broken);
+                _preAttentionErrors = null;
+                _preAttentionWarnings = null;
+            }
+
+            return allErrors;
+        }
+
+        private void AddErrorsToCollection(SqlErrorCollection inCollection, ref SqlErrorCollection collectionToAddTo, ref bool broken)
+        {
+            if (inCollection != null)
+            {
+                foreach (SqlError error in inCollection)
+                {
+                    collectionToAddTo.Add(error);
+                    broken |= (error.Class >= TdsEnums.FATAL_ERROR_CLASS);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Stores away current errors and warnings so that an attention can be processed
+        /// </summary>
+        internal void StoreErrorAndWarningForAttention()
+        {
+            lock (_errorAndWarningsLock)
+            {
+                Debug.Assert(_preAttentionErrors == null && _preAttentionWarnings == null, "Can't store errors for attention because there are already errors stored");
+
+                _hasErrorOrWarning = false;
+
+                _preAttentionErrors = _errors;
+                _preAttentionWarnings = _warnings;
+
+                _errors = null;
+                _warnings = null;
+            }
+        }
+
+        /// <summary>
+        /// Restores errors and warnings that were stored in order to process an attention
+        /// </summary>
+        internal void RestoreErrorAndWarningAfterAttention()
+        {
+            lock (_errorAndWarningsLock)
+            {
+                Debug.Assert(_errors == null && _warnings == null, "Can't restore errors after attention because there are already other errors");
+
+                _hasErrorOrWarning = (((_preAttentionErrors != null) && (_preAttentionErrors.Count > 0)) || ((_preAttentionWarnings != null) && (_preAttentionWarnings.Count > 0)));
+
+                _errors = _preAttentionErrors;
+                _warnings = _preAttentionWarnings;
+
+                _preAttentionErrors = null;
+                _preAttentionWarnings = null;
+            }
+        }
+
+        /// <summary>
+        /// Checks if an error is stored in _error and, if so, throws an error
+        /// </summary>
+        internal void CheckThrowSNIException()
+        {
+            if (HasErrorOrWarning)
+            {
+                ThrowExceptionAndWarning();
+            }
+        }
+
+        private static void ReadAsyncCallbackComplete(object state)
+        {
+            TaskCompletionSource<object> source = (TaskCompletionSource<object>)state;
+            source.TrySetResult(null);
+        }
+
+        /////////////////////////////////////////
+        // Network/Packet Writing & Processing //
+        /////////////////////////////////////////
+
+        //
+        // Takes a secure string and offsets and saves them for a write latter when the information is written out to SNI Packet
+        //  This method is provided to better handle the life cycle of the clear text of the secure string
+        //  This method also ensures that the clear text is not held in the unpined managed buffer so that it avoids getting moved around by CLR garbage collector
+        //  TdsParserStaticMethods.EncryptPassword operation is also done in the unmanaged buffer for the clear text later
+        //
+        internal void WriteSecureString(SecureString secureString)
+        {
+            Debug.Assert(_securePasswords[0] == null || _securePasswords[1] == null, "There are more than two secure passwords");
+
+            int index = _securePasswords[0] != null ? 1 : 0;
+
+            _securePasswords[index] = secureString;
+            _securePasswordOffsetsInBuffer[index] = _outBytesUsed;
+
+            // loop through and write the entire array
+            int lengthInBytes = secureString.Length * 2;
+
+            // It is guaranteed both secure password and secure change password should fit into the first packet
+            // Given current TDS format and implementation it is not possible that one of secure string is the last item and exactly fill up the output buffer
+            //  if this ever happens and it is correct situation, the packet needs to be written out after _outBytesUsed is update
+            Debug.Assert((_outBytesUsed + lengthInBytes) < _outBuff.Length, "Passwords cannot be split into two different packet or the last item which fully fill up _outBuff!!!");
+
+            _outBytesUsed += lengthInBytes;
+        }
+
+        internal void ResetSecurePasswordsInformation()
+        {
+            for (int i = 0; i < _securePasswords.Length; ++i)
+            {
+                _securePasswords[i] = null;
+                _securePasswordOffsetsInBuffer[i] = 0;
+            }
+        }
+
+        internal Task WaitForAccumulatedWrites()
+        {
+            // Checked for stored exceptions
+            Exception delayedException = Interlocked.Exchange(ref _delayedWriteAsyncCallbackException, null);
+            if (delayedException != null)
+            {
+                throw delayedException;
+            }
+#pragma warning restore 420
+
+            if (_asyncWriteCount == 0)
+            {
+                return null;
+            }
+
+            _writeCompletionSource = new TaskCompletionSource<object>();
+            Task task = _writeCompletionSource.Task;
+
+            // Ensure that _writeCompletionSource is set before checking state
+            Interlocked.MemoryBarrier();
+
+            // Now that we have set _writeCompletionSource, check if parser is closed or broken
+            if ((_parser.State == TdsParserState.Closed) || (_parser.State == TdsParserState.Broken))
+            {
+                throw ADP.ClosedConnectionError();
+            }
+
+            // Check for stored exceptions
+            delayedException = Interlocked.Exchange(ref _delayedWriteAsyncCallbackException, null);
+            if (delayedException != null)
+            {
+                throw delayedException;
+            }
+
+            // If there are no outstanding writes, see if we can shortcut and return null
+            if ((_asyncWriteCount == 0) && ((!task.IsCompleted) || (task.Exception == null)))
+            {
+                task = null;
+            }
+
+            return task;
+        }
+
+        // Takes in a single byte and writes it to the buffer.  If the buffer is full, it is flushed
+        // and then the buffer is re-initialized in flush() and then the byte is put in the buffer.
+        internal void WriteByte(byte b)
+        {
+            Debug.Assert(_outBytesUsed <= _outBuff.Length, "ERROR - TDSParser: _outBytesUsed > _outBuff.Length");
+
+            // check to make sure we haven't used the full amount of space available in the buffer, if so, flush it
+            if (_outBytesUsed == _outBuff.Length)
+            {
+                WritePacket(TdsEnums.SOFTFLUSH, canAccumulate: true);
+            }
+            // set byte in buffer and increment the counter for number of bytes used in the out buffer
+            _outBuff[_outBytesUsed++] = b;
+        }
+
+        private void CancelWritePacket()
+        {
+            Debug.Assert(_cancelled, "Should not call CancelWritePacket if _cancelled is not set");
+
+            _parser.Connection.ThreadHasParserLockForClose = true;      // In case of error, let the connection know that we are holding the lock
+            try
+            {
+                // Send the attention and wait for the ATTN_ACK
+                SendAttention();
+                ResetCancelAndProcessAttention();
+
+                // Let the caller know that we've given up
+                throw SQL.OperationCancelled();
+            }
+            finally
+            {
+                _parser.Connection.ThreadHasParserLockForClose = false;
+            }
         }
 
         /////////////////////////////////////////
@@ -2886,17 +3462,17 @@ namespace Microsoft.Data.SqlClient
                 _snapshotStatus == TdsParserStateObject.SnapshotStatus.ContinueRunning;
         }
 
-        internal (bool IsAvailable, bool IsStarting, bool IsContinuing) GetSnapshotStatuses()
+        internal (bool CanContinue, bool IsStarting, bool IsContinuing) GetSnapshotStatuses()
         {
-            bool isAvailable = _snapshot != null && _snapshot.ContinueEnabled;
+            bool canContinue = _snapshot != null && _snapshot.ContinueEnabled && _snapshotStatus != SnapshotStatus.NotActive;
             bool isStarting = false;
             bool isContinuing = false;
-            if (isAvailable)
+            if (canContinue)
             {
                 isStarting = _snapshotStatus == SnapshotStatus.ReplayStarting;
                 isContinuing = _snapshotStatus == SnapshotStatus.ContinueRunning;
             }
-            return (isAvailable, isStarting, isContinuing);
+            return (canContinue, isStarting, isContinuing);
         }
 
         internal int GetSnapshotStorageLength<T>()
@@ -2956,6 +3532,110 @@ namespace Microsoft.Data.SqlClient
         {
             Debug.Assert(_snapshot != null && _snapshot.ContinueEnabled, "_snapshot must exist to read packet data size");
             return _snapshot.GetPacketID();
+        }
+
+        /// <summary>
+        /// Debug Only: Ensures that the TdsParserStateObject has no lingering state and can safely be re-used
+        /// </summary>
+        [Conditional("DEBUG")]
+        internal void AssertStateIsClean()
+        {
+            // If our TdsParser is closed or broken, then we don't really care about our state
+            TdsParser parser = _parser;
+            if ((parser != null) && (parser.State != TdsParserState.Closed) && (parser.State != TdsParserState.Broken))
+            {
+                // Async reads
+                Debug.Assert(_snapshot == null && _snapshotStatus == SnapshotStatus.NotActive, "StateObj has leftover snapshot state");
+                Debug.Assert(!_asyncReadWithoutSnapshot, "StateObj has AsyncReadWithoutSnapshot still enabled");
+                Debug.Assert(_executionContext == null, "StateObj has a stored execution context from an async read");
+                // Async writes
+                Debug.Assert(_asyncWriteCount == 0, "StateObj still has outstanding async writes");
+                Debug.Assert(_delayedWriteAsyncCallbackException == null, "StateObj has an unobserved exceptions from an async write");
+                // Attention\Cancellation\Timeouts
+                Debug.Assert(!HasReceivedAttention && !_attentionSent && !_attentionSending, $"StateObj is still dealing with attention: Sent: {_attentionSent}, Received: {HasReceivedAttention}, Sending: {_attentionSending}");
+                Debug.Assert(!_cancelled, "StateObj still has cancellation set");
+                Debug.Assert(_timeoutState == TimeoutState.Stopped, "StateObj still has internal timeout set");
+                // Errors and Warnings
+                Debug.Assert(!_hasErrorOrWarning, "StateObj still has stored errors or warnings");
+            }
+        }
+
+#if DEBUG
+        internal void CompletePendingReadWithSuccess(bool resetForcePendingReadsToWait)
+        {
+            TaskCompletionSource<object> realNetworkPacketTaskSource = _realNetworkPacketTaskSource;
+            TaskCompletionSource<object> networkPacketTaskSource = _networkPacketTaskSource;
+
+            Debug.Assert(s_forcePendingReadsToWaitForUser, "Not forcing pends to wait for user - can't force complete");
+            Debug.Assert(networkPacketTaskSource != null, "No pending read to complete");
+
+            try
+            {
+                if (realNetworkPacketTaskSource != null)
+                {
+                    // Wait for the real read to complete
+                    realNetworkPacketTaskSource.Task.Wait();
+                }
+            }
+            finally
+            {
+                if (networkPacketTaskSource != null)
+                {
+                    if (resetForcePendingReadsToWait)
+                    {
+                        s_forcePendingReadsToWaitForUser = false;
+                    }
+
+                    networkPacketTaskSource.TrySetResult(null);
+                }
+            }
+        }
+
+        internal void CompletePendingReadWithFailure(int errorCode, bool resetForcePendingReadsToWait)
+        {
+            TaskCompletionSource<object> realNetworkPacketTaskSource = _realNetworkPacketTaskSource;
+            TaskCompletionSource<object> networkPacketTaskSource = _networkPacketTaskSource;
+
+            Debug.Assert(s_forcePendingReadsToWaitForUser, "Not forcing pends to wait for user - can't force complete");
+            Debug.Assert(networkPacketTaskSource != null, "No pending read to complete");
+
+            try
+            {
+                if (realNetworkPacketTaskSource != null)
+                {
+                    // Wait for the real read to complete
+                    realNetworkPacketTaskSource.Task.Wait();
+                }
+            }
+            finally
+            {
+                if (networkPacketTaskSource != null)
+                {
+                    if (resetForcePendingReadsToWait)
+                    {
+                        s_forcePendingReadsToWaitForUser = false;
+                    }
+
+                    AddError(new SqlError(errorCode, 0x00, TdsEnums.FATAL_ERROR_CLASS, _parser.Server, string.Empty, string.Empty, 0));
+                    try
+                    {
+                        ThrowExceptionAndWarning();
+                    }
+                    catch (Exception ex)
+                    {
+                        networkPacketTaskSource.TrySetException(ex);
+                    }
+                }
+            }
+        }
+#endif
+
+        internal void CloneCleanupAltMetaDataSetArray()
+        {
+            if (_snapshot != null)
+            {
+                _snapshot.CloneCleanupAltMetaDataSetArray();
+            }
         }
 
         internal sealed partial class StateSnapshot
