@@ -1913,12 +1913,14 @@ CREATE TABLE {tableName} (id INT, foo VARBINARY(MAX))
 
         private static void TestXEventsStreaming(string connectionString)
         {
-            string sessionName = DataTestUtility.GenerateRandomCharacters("Session");
-
-            try
+            // Create XEvent
+            using (SqlConnection xEventManagementConnection = new SqlConnection(connectionString))
+            using (DataTestUtility.XEventScope xEventScope = new DataTestUtility.XEventScope(xEventManagementConnection,
+                "ADD EVENT sqlserver.user_event(ACTION(package0.event_sequence))",
+                "ADD TARGET package0.ring_buffer"))
             {
-                //Create XEvent
-                SetupXevent(connectionString, sessionName);
+                string sessionName = xEventScope.SessionName;
+
                 Task.Factory.StartNew(() =>
                 {
                     // Read XEvents
@@ -1956,52 +1958,6 @@ CREATE TABLE {tableName} (id INT, foo VARBINARY(MAX))
                         }
                     }
                 }).Wait(10000);
-            }
-            finally
-            {
-                //Delete XEvent 
-                DeleteXevent(connectionString, sessionName);
-            }
-        }
-
-        private static void SetupXevent(string connectionString, string sessionName)
-        {
-            string xEventCreateAndStartCommandText = @"CREATE EVENT SESSION [" + sessionName + @"] ON SERVER
-                        ADD EVENT sqlserver.user_event(ACTION(package0.event_sequence))
-                        ADD TARGET package0.ring_buffer
-                        WITH (
-                            MAX_MEMORY=4096 KB,
-                            EVENT_RETENTION_MODE=ALLOW_SINGLE_EVENT_LOSS,
-                            MAX_DISPATCH_LATENCY=30 SECONDS,
-                            MAX_EVENT_SIZE=0 KB,
-                            MEMORY_PARTITION_MODE=NONE,
-                            TRACK_CAUSALITY=ON,
-                            STARTUP_STATE=OFF)
-                            
-                        ALTER EVENT SESSION [" + sessionName + "] ON SERVER STATE = START ";
-
-            using (SqlConnection connection = new SqlConnection(connectionString))
-            {
-                connection.Open();
-                using (SqlCommand createXeventSession = new SqlCommand(xEventCreateAndStartCommandText, connection))
-                {
-                    createXeventSession.ExecuteNonQuery();
-                }
-            }
-        }
-
-        private static void DeleteXevent(string connectionString, string sessionName)
-        {
-            string deleteXeventSessionCommand = $"IF EXISTS (select * from sys.server_event_sessions where name ='{sessionName}')" +
-                    $" DROP EVENT SESSION [{sessionName}] ON SERVER";
-
-            using (SqlConnection connection = new SqlConnection(connectionString))
-            {
-                connection.Open();
-                using (SqlCommand deleteXeventSession = new SqlCommand(deleteXeventSessionCommand, connection))
-                {
-                    deleteXeventSession.ExecuteNonQuery();
-                }
             }
         }
 
